GH-121 Split versioned_store.rs into focused sub-modules (Phase 2)
Split the 4,721-line src/git/versioned_store.rs into a module directory
with 6 focused files:

- mod.rs: struct defs, traits, type aliases
- core.rs: generic KV operations (insert/get/commit/branch/log)
- history.rs: versioning infra (resolve, read config, merge_generic)
- backends.rs: per-storage init/open/checkout/merge/TreeConfigSaver
- thread_safe.rs: ThreadSafeVersionedKvStore wrapper
- tests.rs: all tests

No behavior changes. Module path crate::git::versioned_store remains
the same so all external imports continue working unchanged.
