--- src/git/versioned_store.rs
@@ -1,4721 +1,0 @@
-/*
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-use crate::config::TreeConfig;
-use crate::diff::{ConflictResolver, IgnoreConflictsResolver};
-use crate::digest::ValueDigest;
-use crate::git::types::*;
-use crate::storage::GitNodeStorage;
-use crate::storage::{FileNodeStorage, InMemoryNodeStorage, NodeStorage};
-use crate::tree::{ProllyTree, Tree};
-use gix::prelude::*;
-use parking_lot::Mutex;
-use std::collections::HashMap;
-use std::path::Path;
-use std::sync::Arc;
-
-/// Trait for accessing historical state from version control
-pub trait HistoricalAccess<const N: usize> {
-    /// Get all key-value pairs at a specific reference (commit, branch, etc.)
-    fn get_keys_at_ref(&self, reference: &str) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError>;
-}
-
-/// Trait for accessing commit history and tracking changes to specific keys
-pub trait TreeConfigSaver<const N: usize> {
-    fn save_tree_config_to_git_internal(&self) -> Result<(), GitKvError>;
-}
-
-pub trait HistoricalCommitAccess<const N: usize> {
-    /// Get all commits that contain changes to a specific key
-    /// Returns commits in reverse chronological order (newest first)
-    fn get_commits_for_key(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError>;
-
-    /// Get the commit history for the repository
-    /// Returns commits in reverse chronological order (newest first)
-    fn get_commit_history(&self) -> Result<Vec<CommitInfo>, GitKvError>;
-}
-
-#[cfg(feature = "rocksdb_storage")]
-use crate::storage::RocksDBNodeStorage;
-
-/// A versioned key-value store backed by Git and ProllyTree with configurable storage
-///
-/// This combines the efficient tree operations of ProllyTree with Git's
-/// version control capabilities, providing a full-featured versioned
-/// key-value store with branching, merging, and history.
-pub struct VersionedKvStore<const N: usize, S: NodeStorage<N>> {
-    tree: ProllyTree<N, S>,
-    git_repo: gix::Repository,
-    staging_area: HashMap<Vec<u8>, Option<Vec<u8>>>, // None = deleted
-    current_branch: String,
-    storage_backend: StorageBackend,
-    /// Dataset directory for storing config and other metadata for all backends
-    /// (File/RocksDB/InMemory/Git). Git init/open paths set this from
-    /// `storage.dataset_dir()`, and commit()/merge rely on this field being `Some`.
-    dataset_dir: Option<std::path::PathBuf>,
-}
-
-/// Type alias for backward compatibility (Git storage)
-pub type GitVersionedKvStore<const N: usize> = VersionedKvStore<N, GitNodeStorage<N>>;
-
-/// Type alias for InMemory storage
-pub type InMemoryVersionedKvStore<const N: usize> = VersionedKvStore<N, InMemoryNodeStorage<N>>;
-
-/// Type alias for File storage
-pub type FileVersionedKvStore<const N: usize> = VersionedKvStore<N, FileNodeStorage<N>>;
-
-/// Type alias for RocksDB storage
-#[cfg(feature = "rocksdb_storage")]
-pub type RocksDBVersionedKvStore<const N: usize> = VersionedKvStore<N, RocksDBNodeStorage<N>>;
-
-/// Thread-safe wrapper for VersionedKvStore
-///
-/// This wrapper provides thread-safe access to the underlying VersionedKvStore by using
-/// `Arc<Mutex<>>` internally. All operations are synchronized, making it safe to use
-/// across multiple threads.
-///
-/// Uses `parking_lot::Mutex` which does not support lock poisoning. If a thread panics
-/// while holding the lock, subsequent lock acquisitions will succeed and the data
-/// remains accessible. This is intentional: the previous `std::sync::Mutex` approach
-/// also panicked (via `.unwrap()`) on poisoned locks, so callers had no recovery path
-/// either way. With `parking_lot`, the application at least has a chance to continue.
-pub struct ThreadSafeVersionedKvStore<const N: usize, S: NodeStorage<N>> {
-    inner: Arc<Mutex<VersionedKvStore<N, S>>>,
-}
-
-/// Type alias for thread-safe Git storage
-pub type ThreadSafeGitVersionedKvStore<const N: usize> =
-    ThreadSafeVersionedKvStore<N, GitNodeStorage<N>>;
-
-/// Type alias for thread-safe InMemory storage
-pub type ThreadSafeInMemoryVersionedKvStore<const N: usize> =
-    ThreadSafeVersionedKvStore<N, InMemoryNodeStorage<N>>;
-
-/// Type alias for thread-safe File storage
-pub type ThreadSafeFileVersionedKvStore<const N: usize> =
-    ThreadSafeVersionedKvStore<N, FileNodeStorage<N>>;
-
-/// Type alias for thread-safe RocksDB storage
-#[cfg(feature = "rocksdb_storage")]
-pub type ThreadSafeRocksDBVersionedKvStore<const N: usize> =
-    ThreadSafeVersionedKvStore<N, RocksDBNodeStorage<N>>;
-
-impl<const N: usize, S: NodeStorage<N>> VersionedKvStore<N, S>
-where
-    Self: TreeConfigSaver<N>,
-{
-    /// Find the git repository root by walking up the directory tree
-    fn find_git_root<P: AsRef<Path>>(start_path: P) -> Option<std::path::PathBuf> {
-        let mut current = start_path.as_ref().to_path_buf();
-        loop {
-            if current.join(".git").exists() {
-                return Some(current);
-            }
-            if !current.pop() {
-                break;
-            }
-        }
-        None
-    }
-
-    /// Get the common git directory path, handling worktrees and submodules.
-    /// For worktrees, this resolves to the main .git directory (not the per-worktree gitdir)
-    /// to ensure shared node storage is placed in a common location.
-    fn resolve_git_dir<P: AsRef<Path>>(git_root: P) -> std::path::PathBuf {
-        let git_path = git_root.as_ref().join(".git");
-
-        // If .git is a file (worktree or submodule), read the gitdir path from it
-        let gitdir = if git_path.is_file() {
-            if let Ok(content) = std::fs::read_to_string(&git_path) {
-                let mut resolved_gitdir = None;
-                for line in content.lines() {
-                    if let Some(gitdir_str) = line.strip_prefix("gitdir:") {
-                        let gitdir_str = gitdir_str.trim();
-                        // Handle both absolute and relative paths
-                        let gitdir_path = std::path::Path::new(gitdir_str);
-                        resolved_gitdir = Some(if gitdir_path.is_absolute() {
-                            gitdir_path.to_path_buf()
-                        } else {
-                            git_root.as_ref().join(gitdir_str)
-                        });
-                        break;
-                    }
-                }
-                resolved_gitdir.unwrap_or(git_path)
-            } else {
-                git_path
-            }
-        } else {
-            // Default: .git is a directory
-            git_path
-        };
-
-        // For linked worktrees, resolve to the common git directory
-        // The commondir file contains the path to the main .git directory
-        let commondir_path = gitdir.join("commondir");
-        if commondir_path.is_file() {
-            if let Ok(content) = std::fs::read_to_string(&commondir_path) {
-                let commondir_str = content.trim();
-                if !commondir_str.is_empty() {
-                    let commondir = std::path::Path::new(commondir_str);
-                    // Handle both absolute and relative paths
-                    let resolved_commondir = if commondir.is_absolute() {
-                        commondir.to_path_buf()
-                    } else {
-                        // Relative path is relative to the gitdir
-                        gitdir.join(commondir)
-                    };
-                    // Canonicalize to resolve .. components
-                    if let Ok(canonical) = resolved_commondir.canonicalize() {
-                        return canonical;
-                    }
-                    return resolved_commondir;
-                }
-            }
-        }
-
-        gitdir
-    }
-
-    /// Get the prolly directory path inside the git directory
-    /// This is where all ProllyTree data is stored to avoid accidental git versioning
-    fn get_prolly_dir<P: AsRef<Path>>(git_root: P) -> std::path::PathBuf {
-        Self::resolve_git_dir(git_root).join("prolly")
-    }
-
-    /// Ensure the prolly directory structure exists
-    fn ensure_prolly_dir<P: AsRef<Path>>(git_root: P) -> Result<std::path::PathBuf, GitKvError> {
-        let prolly_dir = Self::get_prolly_dir(&git_root);
-        std::fs::create_dir_all(&prolly_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create prolly directory: {e}"))
-        })?;
-        Ok(prolly_dir)
-    }
-
-    /// Check if the given path is the git repository root directory
-    /// This is used to prevent initializing a dataset at the git root,
-    /// which could cause `git add -A .` to stage unrelated files.
-    fn is_in_git_root<P: AsRef<Path>>(path: P) -> Result<bool, GitKvError> {
-        let path = path.as_ref();
-
-        // Try to canonicalize the path. If it doesn't exist, use the parent directory.
-        let canonical_path = if path.exists() {
-            path.canonicalize()
-                .map_err(|e| GitKvError::GitObjectError(format!("Failed to resolve path: {e}")))?
-        } else {
-            // Path doesn't exist yet (common for init). Use parent + last component.
-            let parent = path.parent().ok_or_else(|| {
-                GitKvError::GitObjectError("Invalid path: no parent directory".to_string())
-            })?;
-
-            // If parent doesn't exist either, we can't proceed
-            if !parent.exists() && !parent.as_os_str().is_empty() {
-                return Err(GitKvError::GitObjectError(format!(
-                    "Parent directory does not exist: {}",
-                    parent.display()
-                )));
-            }
-
-            // Canonicalize parent and append the last component
-            let canonical_parent = if parent.as_os_str().is_empty() {
-                std::env::current_dir().map_err(|e| {
-                    GitKvError::GitObjectError(format!("Failed to get current directory: {e}"))
-                })?
-            } else {
-                parent.canonicalize().map_err(|e| {
-                    GitKvError::GitObjectError(format!("Failed to resolve parent path: {e}"))
-                })?
-            };
-
-            // Append the file name to get the full path
-            if let Some(file_name) = path.file_name() {
-                canonical_parent.join(file_name)
-            } else {
-                canonical_parent
-            }
-        };
-
-        // Find git root from the path (or its parent if path doesn't exist)
-        let lookup_path = if path.exists() {
-            canonical_path.clone()
-        } else {
-            canonical_path
-                .parent()
-                .map(|p| p.to_path_buf())
-                .unwrap_or(canonical_path.clone())
-        };
-
-        if let Some(git_root) = Self::find_git_root(&lookup_path) {
-            let git_root = git_root.canonicalize().map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to resolve git root: {e}"))
-            })?;
-            Ok(canonical_path == git_root)
-        } else {
-            Err(GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            ))
-        }
-    }
-
-    /// Insert a key-value pair (stages the change)
-    pub fn insert(&mut self, key: Vec<u8>, value: Vec<u8>) -> Result<(), GitKvError> {
-        self.staging_area.insert(key, Some(value));
-        self.save_staging_area()?;
-        Ok(())
-    }
-
-    /// Update an existing key-value pair (stages the change)
-    pub fn update(&mut self, key: Vec<u8>, value: Vec<u8>) -> Result<bool, GitKvError> {
-        let exists = self.get(&key).is_some();
-        if exists {
-            self.staging_area.insert(key, Some(value));
-            self.save_staging_area()?;
-        }
-        Ok(exists)
-    }
-
-    /// Delete a key-value pair (stages the change)
-    pub fn delete(&mut self, key: &[u8]) -> Result<bool, GitKvError> {
-        let exists = self.get(key).is_some();
-        if exists {
-            self.staging_area.insert(key.to_vec(), None);
-            self.save_staging_area()?;
-        }
-        Ok(exists)
-    }
-
-    /// Get a value by key (checks staging area first, then committed data)
-    pub fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
-        // Check staging area first
-        if let Some(staged_value) = self.staging_area.get(key) {
-            return staged_value.clone();
-        }
-
-        // Check committed data
-        self.tree.find(key).and_then(|node| {
-            // Find the value in the node
-            node.keys
-                .iter()
-                .position(|k| k == key)
-                .map(|index| node.values[index].clone())
-        })
-    }
-
-    /// List all keys (includes staged changes)
-    pub fn list_keys(&self) -> Vec<Vec<u8>> {
-        let mut keys = std::collections::HashSet::new();
-
-        // Add keys from the committed ProllyTree
-        for key in self.tree.collect_keys() {
-            keys.insert(key);
-        }
-
-        // Add keys from staging area (overrides committed data)
-        for (key, value) in &self.staging_area {
-            if value.is_some() {
-                keys.insert(key.clone());
-            } else {
-                keys.remove(key);
-            }
-        }
-
-        keys.into_iter().collect()
-    }
-
-    /// Show current staging area status
-    pub fn status(&self) -> Vec<(Vec<u8>, String)> {
-        let mut status = Vec::new();
-
-        for (key, value) in &self.staging_area {
-            let status_str = match value {
-                Some(_) => {
-                    if self.tree.find(key).is_some() {
-                        "modified".to_string()
-                    } else {
-                        "added".to_string()
-                    }
-                }
-                None => "deleted".to_string(),
-            };
-            status.push((key.clone(), status_str));
-        }
-
-        status
-    }
-
-    /// Commit staged changes
-    pub fn commit(&mut self, message: &str) -> Result<gix::ObjectId, GitKvError> {
-        // Apply staged changes to the tree
-        for (key, value) in self.staging_area.drain() {
-            match value {
-                Some(v) => {
-                    self.tree.insert(key, v);
-                }
-                None => {
-                    self.tree.delete(&key);
-                }
-            }
-        }
-
-        // Persist the tree state (including updating root hash and saving config)
-        self.tree.persist_root();
-
-        // For all storage types, also save the tree config to git for historical access
-        self.save_tree_config_to_git_internal()?;
-
-        // Create tree object in Git using git commands
-        // Get the git root directory using work_dir() for worktree/submodule compatibility
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".into()))?;
-        let git_root = self
-            .git_repo
-            .work_dir()
-            .map(|p| p.to_path_buf())
-            .or_else(|| Self::find_git_root(dataset_dir))
-            .ok_or_else(|| GitKvError::GitObjectError("Could not find git root".into()))?;
-
-        // Stage all files in the current directory recursively
-        let add_cmd = std::process::Command::new("git")
-            .args(["add", "-A", "."])
-            .current_dir(&git_root)
-            .output()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to run git add: {e}")))?;
-
-        if !add_cmd.status.success() {
-            let stderr = String::from_utf8_lossy(&add_cmd.stderr);
-            eprintln!("Warning: git add failed: {stderr}");
-        }
-
-        // Use git write-tree to create tree from the current index
-        let write_tree_cmd = std::process::Command::new("git")
-            .args(["write-tree"])
-            .current_dir(&git_root)
-            .output()
-            .map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to run git write-tree: {e}"))
-            })?;
-
-        if !write_tree_cmd.status.success() {
-            let stderr = String::from_utf8_lossy(&write_tree_cmd.stderr);
-            return Err(GitKvError::GitObjectError(format!(
-                "git write-tree failed: {stderr}"
-            )));
-        }
-
-        let tree_hash = String::from_utf8_lossy(&write_tree_cmd.stdout)
-            .trim()
-            .to_string();
-        let tree_id = gix::ObjectId::from_hex(tree_hash.as_bytes())
-            .map_err(|e| GitKvError::GitObjectError(format!("Invalid tree hash: {e}")))?;
-
-        // Create commit
-        let commit_id = self.create_git_commit(tree_id, message)?;
-
-        // Update HEAD
-        self.update_head(commit_id)?;
-
-        // Clear staging area file since we've committed
-        self.save_staging_area()?;
-
-        Ok(commit_id)
-    }
-
-    /// Create a new branch
-    pub fn branch(&mut self, name: &str) -> Result<(), GitKvError> {
-        // Get the current HEAD commit
-        let head = self
-            .git_repo
-            .head()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get HEAD: {e}")))?;
-
-        let head_commit_id = head.id().ok_or_else(|| {
-            GitKvError::GitObjectError("HEAD does not point to a commit".to_string())
-        })?;
-
-        // Create the branch reference to point to the current HEAD
-        let refs_dir = self.git_repo.path().join("refs").join("heads");
-        std::fs::create_dir_all(&refs_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create refs directory: {e}"))
-        })?;
-
-        let branch_file = refs_dir.join(name);
-
-        // Create parent directories if the branch name contains slashes
-        if let Some(parent) = branch_file.parent() {
-            std::fs::create_dir_all(parent).map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to create branch directory: {e}"))
-            })?;
-        }
-
-        std::fs::write(&branch_file, head_commit_id.to_hex().to_string()).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to write branch reference: {e}"))
-        })?;
-
-        Ok(())
-    }
-
-    /// Create a new branch from the current branch and switch to it
-    pub fn create_branch(&mut self, name: &str) -> Result<(), GitKvError> {
-        // First create the branch
-        self.branch(name)?;
-
-        // Then switch to it
-        // Clear staging area
-        self.staging_area.clear();
-        self.save_staging_area()?;
-
-        // Update our internal tracking to the new branch
-        self.current_branch = name.to_string();
-
-        // Update HEAD to point to the new branch
-        let head_file = self.git_repo.path().join("HEAD");
-        let head_content = format!("ref: refs/heads/{name}");
-        std::fs::write(&head_file, head_content)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to update HEAD: {e}")))?;
-
-        // Note: Tree reload is handled in Git-specific implementation
-
-        Ok(())
-    }
-
-    // Note: checkout is implemented differently for each storage type
-    // GitNodeStorage has its own implementation that reloads tree state
-
-    /// Get current branch name
-    pub fn current_branch(&self) -> &str {
-        &self.current_branch
-    }
-
-    /// List all branches in the repository
-    pub fn list_branches(&self) -> Result<Vec<String>, GitKvError> {
-        let mut branches = Vec::new();
-
-        // Get all refs under refs/heads/
-        let refs = self
-            .git_repo
-            .refs
-            .iter()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to iterate refs: {e}")))?;
-
-        for reference in (refs
-            .all()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get refs: {e}")))?)
-        .flatten()
-        {
-            if let Some(name) = reference.name.as_bstr().strip_prefix(b"refs/heads/") {
-                let branch_name = String::from_utf8_lossy(name).to_string();
-                branches.push(branch_name);
-            }
-        }
-
-        branches.sort();
-        Ok(branches)
-    }
-
-    /// Get access to the git repository (for internal use)
-    pub fn git_repo(&self) -> &gix::Repository {
-        &self.git_repo
-    }
-
-    /// Get commit history
-    pub fn log(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        let mut history = Vec::new();
-
-        // Get the current HEAD commit
-        let head_commit = match self.git_repo.head_commit() {
-            Ok(commit) => commit,
-            Err(_) => return Ok(history), // No commits yet
-        };
-
-        // Use rev_walk to traverse the commit history
-        let rev_walk = self.git_repo.rev_walk([head_commit.id()]);
-
-        match rev_walk.all() {
-            Ok(walk) => {
-                for info in walk.take(100).flatten() {
-                    // Limit to 100 commits
-                    if let Ok(commit_obj) = info.object() {
-                        if let Ok(commit_ref) = commit_obj.decode() {
-                            let commit_info = CommitInfo {
-                                id: commit_obj.id().into(),
-                                author: format!(
-                                    "{} <{}>",
-                                    String::from_utf8_lossy(commit_ref.author.name),
-                                    String::from_utf8_lossy(commit_ref.author.email)
-                                ),
-                                committer: format!(
-                                    "{} <{}>",
-                                    String::from_utf8_lossy(commit_ref.committer.name),
-                                    String::from_utf8_lossy(commit_ref.committer.email)
-                                ),
-                                message: String::from_utf8_lossy(commit_ref.message).to_string(),
-                                timestamp: commit_ref.author.time.seconds,
-                            };
-                            history.push(commit_info);
-                        }
-                    }
-                }
-            }
-            Err(_) => {
-                // Fallback to single commit if rev_walk fails
-                let commit_info = CommitInfo {
-                    id: head_commit.id().into(),
-                    author: "Unknown".to_string(),
-                    committer: "Unknown".to_string(),
-                    message: "Commit".to_string(),
-                    timestamp: 0,
-                };
-                history.push(commit_info);
-            }
-        }
-
-        Ok(history)
-    }
-
-    /// Get git user configuration (name and email)
-    fn get_git_user_config(&self) -> Result<(String, String), GitKvError> {
-        let config = self.git_repo.config_snapshot();
-
-        let name = config
-            .string("user.name")
-            .map(|n| n.to_string())
-            .unwrap_or_else(|| "git-prolly".to_string());
-
-        let email = config
-            .string("user.email")
-            .map(|e| e.to_string())
-            .unwrap_or_else(|| "git-prolly@example.com".to_string());
-
-        Ok((name, email))
-    }
-
-    /// Create a Git commit object
-    fn create_git_commit(
-        &self,
-        tree_id: gix::ObjectId,
-        message: &str,
-    ) -> Result<gix::ObjectId, GitKvError> {
-        // Get the current time
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map_err(|e| GitKvError::GitObjectError(format!("System time error: {e}")))?
-            .as_secs() as i64;
-
-        // Get git user configuration
-        let (name, email) = self.get_git_user_config()?;
-
-        // Create author and committer signatures
-        let signature = gix::actor::Signature {
-            name: name.into(),
-            email: email.into(),
-            time: gix::date::Time {
-                seconds: now,
-                offset: 0,
-                sign: gix::date::time::Sign::Plus,
-            },
-        };
-
-        // Get parent commits (current HEAD if exists)
-        let parent_ids = match self.git_repo.head_commit() {
-            Ok(parent) => vec![parent.id().into()],
-            Err(_) => vec![], // No parent for initial commit
-        };
-
-        // Create commit object
-        let commit = gix::objs::Commit {
-            tree: tree_id,
-            parents: parent_ids.into(),
-            author: signature.clone(),
-            committer: signature,
-            encoding: None,
-            message: message.as_bytes().into(),
-            extra_headers: vec![],
-        };
-
-        let commit_id = self
-            .git_repo
-            .objects
-            .write(&commit)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to write commit: {e}")))?;
-
-        Ok(commit_id)
-    }
-
-    /// Update HEAD to point to the new commit
-    fn update_head(&mut self, commit_id: gix::ObjectId) -> Result<(), GitKvError> {
-        // Update the current branch reference to point to the new commit
-        let branch_ref = format!("refs/heads/{}", self.current_branch);
-
-        // For now, use a simple implementation that writes directly to the file
-        let refs_dir = self.git_repo.path().join("refs").join("heads");
-        std::fs::create_dir_all(&refs_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create refs directory: {e}"))
-        })?;
-
-        let branch_file = refs_dir.join(&self.current_branch);
-        std::fs::write(&branch_file, commit_id.to_hex().to_string()).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to write branch reference: {e}"))
-        })?;
-
-        // Update HEAD to point to the branch
-        let head_file = self.git_repo.path().join("HEAD");
-        std::fs::write(&head_file, format!("ref: {branch_ref}\n")).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to write HEAD reference: {e}"))
-        })?;
-
-        Ok(())
-    }
-
-    /// Save the staging area to a file
-    fn save_staging_area(&self) -> Result<(), GitKvError> {
-        let staging_file = self.get_staging_file_path()?;
-
-        // Serialize the staging area
-        let serialized =
-            bincode::serialize(&self.staging_area).map_err(GitKvError::SerializationError)?;
-
-        std::fs::write(staging_file, serialized).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to write staging area: {e}"))
-        })?;
-
-        Ok(())
-    }
-
-    /// Load the staging area from a file
-    fn load_staging_area(&mut self) -> Result<(), GitKvError> {
-        let staging_file = self.get_staging_file_path()?;
-
-        if staging_file.exists() {
-            let data = std::fs::read(staging_file).map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to read staging area: {e}"))
-            })?;
-
-            self.staging_area =
-                bincode::deserialize(&data).map_err(GitKvError::SerializationError)?;
-        }
-
-        Ok(())
-    }
-
-    /// Get the dataset-specific staging file path
-    fn get_staging_file_path(&self) -> Result<std::path::PathBuf, GitKvError> {
-        // Get the current directory relative to git root
-        let current_dir = std::env::current_dir().map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to get current directory: {e}"))
-        })?;
-
-        let git_root = Self::find_git_root(&current_dir)
-            .ok_or_else(|| GitKvError::GitObjectError("Not in a git repository".to_string()))?;
-
-        // Create a dataset-specific identifier from the relative path
-        let relative_path = current_dir
-            .strip_prefix(&git_root)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get relative path: {e}")))?;
-
-        // Use the relative path to create a unique staging file name
-        let path_str = relative_path.to_string_lossy().replace(['/', '\\'], "_");
-        let staging_filename = if path_str.is_empty() {
-            "PROLLY_STAGING_root".to_string()
-        } else {
-            format!("PROLLY_STAGING_{path_str}")
-        };
-
-        Ok(self.git_repo.path().join(staging_filename))
-    }
-
-    /// Generate a cryptographic proof for a key's existence and value in the tree
-    /// This proof can be used to verify the integrity of the key-value pair without
-    /// requiring access to the entire tree structure.
-    ///
-    /// # Parameters
-    /// - `key`: The key to generate proof for
-    ///
-    /// # Returns
-    /// - A proof object containing the hash path from root to the target node
-    pub fn generate_proof(&self, key: &[u8]) -> crate::proof::Proof<N> {
-        self.tree.generate_proof(key)
-    }
-
-    /// Verify a cryptographic proof for a key-value pair
-    /// This checks that the proof is valid and optionally verifies the expected value
-    ///
-    /// # Parameters
-    /// - `proof`: The proof to verify
-    /// - `key`: The key that the proof claims to prove
-    /// - `expected_value`: Optional expected value to verify against
-    ///
-    /// # Returns
-    /// - `true` if the proof is valid, `false` otherwise
-    pub fn verify(
-        &self,
-        proof: crate::proof::Proof<N>,
-        key: &[u8],
-        expected_value: Option<&[u8]>,
-    ) -> bool {
-        self.tree.verify(proof, key, expected_value)
-    }
-}
-
-#[cfg(test)]
-mod proof_tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_versioned_store_proof_methods() {
-        // Create a temporary directory for the test
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let repo_path = temp_dir.path().to_str().unwrap();
-
-        // Initialize git repo
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to initialize git repo");
-
-        // Set git config
-        std::process::Command::new("git")
-            .args(["config", "user.name", "Test User"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to set git user name");
-
-        std::process::Command::new("git")
-            .args(["config", "user.email", "test@example.com"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to set git user email");
-
-        // Create a subdirectory for the dataset (git-prolly requires this)
-        let dataset_path = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_path).expect("Failed to create dataset directory");
-
-        // Change to the dataset subdirectory
-        let original_dir = std::env::current_dir().expect("Failed to get current dir");
-        std::env::set_current_dir(&dataset_path).expect("Failed to change directory");
-
-        // Initialize the versioned store from the dataset subdirectory
-        let mut store =
-            GitVersionedKvStore::<32>::init(&dataset_path).expect("Failed to initialize store");
-
-        // Insert test data
-        let key = b"proof_test_key".to_vec();
-        let value = b"proof_test_value".to_vec();
-
-        store
-            .insert(key.clone(), value.clone())
-            .expect("Failed to insert");
-        store
-            .commit("Add test data for proof")
-            .expect("Failed to commit");
-
-        // Test generate_proof method exists and works
-        let proof = store.generate_proof(&key);
-
-        // Test verify method with correct value
-        assert!(store.verify(proof.clone(), &key, Some(&value)));
-
-        // Test verify method for existence only
-        assert!(store.verify(proof.clone(), &key, None));
-
-        // Test verify with wrong value should fail
-        let wrong_value = b"wrong_value".to_vec();
-        assert!(!store.verify(proof.clone(), &key, Some(&wrong_value)));
-
-        // Restore original directory (best effort - may fail in parallel tests)
-        let _ = std::env::set_current_dir(original_dir);
-    }
-
-    #[test]
-    fn test_get_keys_at_ref() {
-        // Create a temporary directory for the test
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let repo_path = temp_dir.path().to_str().unwrap();
-
-        // Initialize git repo
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to initialize git repo");
-
-        // Set git config
-        std::process::Command::new("git")
-            .args(["config", "user.name", "Test User"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to set git user name");
-
-        std::process::Command::new("git")
-            .args(["config", "user.email", "test@example.com"])
-            .current_dir(repo_path)
-            .output()
-            .expect("Failed to set git user email");
-
-        // Create a subdirectory for the dataset (git-prolly requires this)
-        let dataset_path = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_path).expect("Failed to create dataset directory");
-
-        // Change to the dataset subdirectory
-        let original_dir = std::env::current_dir().expect("Failed to get current dir");
-        std::env::set_current_dir(&dataset_path).expect("Failed to change directory");
-
-        // Initialize the versioned store from the dataset subdirectory
-        let mut store =
-            GitVersionedKvStore::<32>::init(&dataset_path).expect("Failed to initialize store");
-
-        // Add initial data and commit
-        store
-            .insert(b"key1".to_vec(), b"value1".to_vec())
-            .expect("Failed to insert key1");
-        store
-            .insert(b"key2".to_vec(), b"value2".to_vec())
-            .expect("Failed to insert key2");
-        let commit1 = store.commit("Initial commit").expect("Failed to commit");
-
-        // Get keys at HEAD (should have key1 and key2)
-        let keys_at_head = store
-            .get_keys_at_ref("HEAD")
-            .expect("Failed to get keys at HEAD");
-        assert_eq!(keys_at_head.len(), 2);
-        assert_eq!(
-            keys_at_head.get(&b"key1".to_vec()),
-            Some(&b"value1".to_vec())
-        );
-        assert_eq!(
-            keys_at_head.get(&b"key2".to_vec()),
-            Some(&b"value2".to_vec())
-        );
-
-        // Add more data and commit
-        store
-            .insert(b"key3".to_vec(), b"value3".to_vec())
-            .expect("Failed to insert key3");
-        store
-            .update(b"key1".to_vec(), b"updated1".to_vec())
-            .expect("Failed to update key1");
-        let _commit2 = store.commit("Second commit").expect("Failed to commit");
-
-        // Get keys at the first commit
-        let keys_at_commit1 = store
-            .get_keys_at_ref(&commit1.to_hex().to_string())
-            .expect("Failed to get keys at commit1");
-        assert_eq!(keys_at_commit1.len(), 2);
-        assert_eq!(
-            keys_at_commit1.get(&b"key1".to_vec()),
-            Some(&b"value1".to_vec())
-        );
-        assert_eq!(
-            keys_at_commit1.get(&b"key2".to_vec()),
-            Some(&b"value2".to_vec())
-        );
-        assert!(!keys_at_commit1.contains_key(&b"key3".to_vec()));
-
-        // Get keys at HEAD~1 (should be same as first commit)
-        // Note: HEAD~1 syntax might not work with gix library, use commit hash instead
-        // let keys_at_head_minus_1 = store
-        //     .get_keys_at_ref("HEAD~1")
-        //     .expect("Failed to get keys at HEAD~1");
-        // assert_eq!(keys_at_head_minus_1, keys_at_commit1);
-
-        // Get keys at current HEAD (should have all three keys with updated key1)
-        let keys_at_current_head = store
-            .get_keys_at_ref("HEAD")
-            .expect("Failed to get keys at current HEAD");
-        assert_eq!(keys_at_current_head.len(), 3);
-        assert_eq!(
-            keys_at_current_head.get(&b"key1".to_vec()),
-            Some(&b"updated1".to_vec())
-        );
-        assert_eq!(
-            keys_at_current_head.get(&b"key2".to_vec()),
-            Some(&b"value2".to_vec())
-        );
-        assert_eq!(
-            keys_at_current_head.get(&b"key3".to_vec()),
-            Some(&b"value3".to_vec())
-        );
-
-        // Restore original directory (best effort - may fail in parallel tests)
-        let _ = std::env::set_current_dir(original_dir);
-    }
-}
-
-// Generic diff functionality for all storage types
-impl<const N: usize, S: NodeStorage<N>> VersionedKvStore<N, S>
-where
-    VersionedKvStore<N, S>: HistoricalAccess<N>,
-{
-    /// Compare two commits or branches and return all keys that are added, updated or deleted
-    pub fn diff(&self, from: &str, to: &str) -> Result<Vec<KvDiff>, GitKvError> {
-        // Get all keys from both references
-        let from_keys = self.get_keys_at_ref(from)?;
-        let to_keys = self.get_keys_at_ref(to)?;
-
-        let mut diffs = Vec::new();
-
-        // Check for added or modified keys
-        for (key, to_value) in &to_keys {
-            match from_keys.get(key) {
-                None => {
-                    // Key was added
-                    diffs.push(KvDiff {
-                        key: key.clone(),
-                        operation: DiffOperation::Added(to_value.clone()),
-                    });
-                }
-                Some(from_value) => {
-                    if from_value != to_value {
-                        // Key was modified
-                        diffs.push(KvDiff {
-                            key: key.clone(),
-                            operation: DiffOperation::Modified {
-                                old: from_value.clone(),
-                                new: to_value.clone(),
-                            },
-                        });
-                    }
-                }
-            }
-        }
-
-        // Check for removed keys
-        for (key, from_value) in &from_keys {
-            if !to_keys.contains_key(key) {
-                diffs.push(KvDiff {
-                    key: key.clone(),
-                    operation: DiffOperation::Removed(from_value.clone()),
-                });
-            }
-        }
-
-        // Sort diffs by key for consistent output
-        diffs.sort_by(|a, b| a.key.cmp(&b.key));
-
-        Ok(diffs)
-    }
-}
-
-// Generic commit history functionality for all storage types
-impl<const N: usize, S: NodeStorage<N>> VersionedKvStore<N, S>
-where
-    VersionedKvStore<N, S>: HistoricalCommitAccess<N>,
-{
-    /// Get all commits that contain changes to a specific key
-    /// Returns commits in reverse chronological order (newest first), similar to `git log -- <file>`
-    pub fn get_commits(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commits_for_key(key)
-    }
-
-    /// Get the current HEAD commit ID
-    pub fn current_commit(&self) -> Result<gix::ObjectId, GitKvError> {
-        let head = self
-            .git_repo
-            .head()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get HEAD: {e}")))?;
-
-        let head_commit_id = head.id().ok_or_else(|| {
-            GitKvError::GitObjectError("HEAD does not point to a commit".to_string())
-        })?;
-
-        Ok(head_commit_id.detach())
-    }
-}
-
-// Implement TreeConfigSaver for GitNodeStorage
-impl<const N: usize> TreeConfigSaver<N> for VersionedKvStore<N, GitNodeStorage<N>> {
-    fn save_tree_config_to_git_internal(&self) -> Result<(), GitKvError> {
-        self.save_tree_config_to_git()
-    }
-}
-
-// Storage-specific implementations
-impl<const N: usize> VersionedKvStore<N, GitNodeStorage<N>> {
-    /// Save both tree config and hash mappings to git for GitNodeStorage
-    fn save_tree_config_to_git(&self) -> Result<(), GitKvError> {
-        // For GitNodeStorage, we need to ensure the config and hash mappings are
-        // available in the dataset directory so they can be committed to git
-
-        // Get the current tree configuration
-        let config = self.tree.config.clone();
-
-        // Serialize the configuration to JSON
-        let config_json = serde_json::to_string_pretty(&config)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to serialize config: {e}")))?;
-
-        // Write config to the dataset directory
-        let config_path = self
-            .tree
-            .storage
-            .dataset_dir()
-            .join("prolly_config_tree_config");
-        std::fs::write(&config_path, config_json)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to write config file: {e}")))?;
-
-        // Get hash mappings from storage and save them
-        let mappings = self.tree.storage.get_hash_mappings();
-        let mut mappings_content = String::new();
-        for (hash, object_id) in mappings {
-            // Convert hash bytes to hex manually
-            let hash_hex: String = hash.as_bytes().iter().map(|b| format!("{b:02x}")).collect();
-            mappings_content.push_str(&format!("{hash_hex}:{object_id}\n"));
-        }
-
-        // Write mappings to the dataset directory
-        let mappings_path = self.tree.storage.dataset_dir().join("prolly_hash_mappings");
-        std::fs::write(&mappings_path, mappings_content).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to write mappings file: {e}"))
-        })?;
-
-        Ok(())
-    }
-
-    /// Git-specific checkout that reloads tree state from the target commit
-    pub fn checkout(&mut self, branch_or_commit: &str) -> Result<(), GitKvError> {
-        // Call the generic checkout to handle HEAD reference update
-        // Clear staging area
-        self.staging_area.clear();
-        self.save_staging_area()?;
-
-        // Update HEAD to point to the new branch/commit
-        let target_ref = if branch_or_commit.starts_with("refs/") {
-            branch_or_commit.to_string()
-        } else {
-            format!("refs/heads/{branch_or_commit}")
-        };
-
-        // Check if the reference exists
-        match self.git_repo.refs.find(&target_ref) {
-            Ok(_reference) => {
-                // Update our internal tracking
-                self.current_branch = branch_or_commit.to_string();
-
-                // Update HEAD to point to the new branch
-                let head_file = self.git_repo.path().join("HEAD");
-                let head_content = format!("ref: refs/heads/{branch_or_commit}");
-                std::fs::write(&head_file, head_content).map_err(|e| {
-                    GitKvError::GitObjectError(format!("Failed to update HEAD: {e}"))
-                })?;
-            }
-            Err(_) => {
-                return Err(GitKvError::BranchNotFound(branch_or_commit.to_string()));
-            }
-        }
-
-        // Git-specific: Reload the tree from the HEAD commit of the target branch
-        self.reload_tree_from_head()?;
-
-        Ok(())
-    }
-
-    /// Merge another branch into the current branch with configurable conflict resolution
-    ///
-    /// This method performs a three-way merge using the ProllyTree merge functionality.
-    /// It merges changes from the source branch into the current (destination) branch.
-    ///
-    /// # Arguments
-    ///
-    /// * `source_branch` - The branch to merge from
-    /// * `resolver` - Conflict resolver to handle merge conflicts
-    ///
-    /// # Returns
-    ///
-    /// Returns Ok(commit_id) if merge succeeded, or Err with unresolved conflicts
-    pub fn merge<R: ConflictResolver>(
-        &mut self,
-        source_branch: &str,
-        resolver: &R,
-    ) -> Result<gix::ObjectId, GitKvError> {
-        // Get current branch name
-        let dest_branch = self.current_branch.clone();
-
-        // Find common base commit
-        let base_commit = self.find_merge_base(&dest_branch, source_branch)?;
-
-        // Note: We don't need the tree root hashes for key-value level merge
-        // let base_root = self.get_tree_root_at_commit(&base_commit)?;
-        // let source_root = self.get_tree_root_at_branch(source_branch)?;
-        // let dest_root = self.tree.get_root_hash().unwrap(); // Current tree
-
-        // Perform the merge at the key-value level instead of tree level
-        // Get the actual key-value data from each branch
-        let base_kv = self.collect_keys_at_commit(&base_commit)?;
-        let source_kv = self.collect_keys_at_commit(&self.get_branch_commit(source_branch)?)?;
-        let mut dest_kv = HashMap::new();
-
-        // Collect current tree data
-        for key in self.tree.collect_keys() {
-            if let Some(value) = self.get(&key) {
-                dest_kv.insert(key, value);
-            }
-        }
-
-        // Perform three-way merge at key-value level
-        let mut merge_results = Vec::new();
-        let mut all_keys = std::collections::HashSet::new();
-
-        // Collect all keys from all three states
-        all_keys.extend(base_kv.keys().cloned());
-        all_keys.extend(source_kv.keys().cloned());
-        all_keys.extend(dest_kv.keys().cloned());
-
-        for key in all_keys {
-            let base_value = base_kv.get(&key);
-            let source_value = source_kv.get(&key);
-            let dest_value = dest_kv.get(&key);
-
-            match (base_value, source_value, dest_value) {
-                // Key exists in all three - check for modifications
-                (Some(base), Some(source), Some(dest)) => {
-                    if base == source && base == dest {
-                        // No changes, skip
-                        continue;
-                    } else if base == dest && base != source {
-                        // Only source changed - take source value
-                        merge_results.push(crate::diff::MergeResult::Modified(key, source.clone()));
-                    } else if base == source && base != dest {
-                        // Only dest changed - keep dest value (no-op)
-                        continue;
-                    } else if source == dest {
-                        // Both changed to same value - keep it (no-op)
-                        continue;
-                    } else {
-                        // Conflict: both branches modified differently
-                        let conflict = crate::diff::MergeConflict {
-                            key: key.clone(),
-                            base_value: Some(base.clone()),
-                            source_value: Some(source.clone()),
-                            destination_value: Some(dest.clone()),
-                        };
-                        merge_results.push(crate::diff::MergeResult::Conflict(conflict));
-                    }
-                }
-                // Key added in source, not in base or dest
-                (None, Some(source), None) => {
-                    merge_results.push(crate::diff::MergeResult::Added(key, source.clone()));
-                }
-                // Key added in dest, not in base or source - keep it (no-op)
-                (None, None, Some(_dest)) => {
-                    continue;
-                }
-                // Key added in both source and dest - potential conflict
-                (None, Some(source), Some(dest)) => {
-                    if source == dest {
-                        // Both added same value - keep it (no-op)
-                        continue;
-                    } else {
-                        // Conflict: both branches added different values
-                        let conflict = crate::diff::MergeConflict {
-                            key: key.clone(),
-                            base_value: None,
-                            source_value: Some(source.clone()),
-                            destination_value: Some(dest.clone()),
-                        };
-                        merge_results.push(crate::diff::MergeResult::Conflict(conflict));
-                    }
-                }
-                // Key deleted in source, still exists in dest
-                (Some(_base), None, Some(_dest)) => {
-                    // Source deleted it - apply deletion
-                    merge_results.push(crate::diff::MergeResult::Removed(key));
-                }
-                // Key deleted in dest, still exists in source - keep deletion (no-op)
-                (Some(_base), Some(_source), None) => {
-                    continue;
-                }
-                // Key deleted in both - no-op
-                (Some(_base), None, None) => {
-                    continue;
-                }
-                // All other cases - no action needed
-                _ => continue,
-            }
-        }
-
-        // Apply conflict resolution
-        let mut resolved_results = Vec::new();
-        let mut unresolved_conflicts = Vec::new();
-
-        for result in merge_results {
-            match result {
-                crate::diff::MergeResult::Conflict(conflict) => {
-                    if let Some(resolved_result) = resolver.resolve_conflict(&conflict) {
-                        resolved_results.push(resolved_result);
-                    } else {
-                        unresolved_conflicts.push(conflict);
-                    }
-                }
-                other => resolved_results.push(other),
-            }
-        }
-
-        // If there are unresolved conflicts, return them
-        if !unresolved_conflicts.is_empty() {
-            return Err(GitKvError::MergeConflictError(unresolved_conflicts));
-        }
-
-        // Apply resolved merge results directly to current tree
-        for result in resolved_results {
-            match result {
-                crate::diff::MergeResult::Added(key, value) => {
-                    self.tree.insert(key, value);
-                }
-                crate::diff::MergeResult::Modified(key, value) => {
-                    self.tree.insert(key, value); // insert overwrites existing
-                }
-                crate::diff::MergeResult::Removed(key) => {
-                    self.tree.delete(&key);
-                }
-                crate::diff::MergeResult::Conflict(_) => {
-                    // Should not happen since we handled conflicts above
-                    unreachable!("Conflicts should have been resolved");
-                }
-            }
-        }
-
-        // Clear staging area since we've applied changes directly to tree
-        self.staging_area.clear();
-        self.save_staging_area()?;
-
-        // Create merge commit
-        let message = format!("Merge branch '{source_branch}' into '{dest_branch}'");
-        let merge_commit_id = self.create_merge_commit(&message, source_branch)?;
-
-        Ok(merge_commit_id)
-    }
-
-    /// Convenience method to merge with default IgnoreConflictsResolver
-    pub fn merge_ignore_conflicts(
-        &mut self,
-        source_branch: &str,
-    ) -> Result<gix::ObjectId, GitKvError> {
-        self.merge(source_branch, &IgnoreConflictsResolver)
-    }
-
-    /// Find the merge base (common ancestor) of two branches
-    fn find_merge_base(&self, branch1: &str, branch2: &str) -> Result<gix::ObjectId, GitKvError> {
-        // Get commit IDs for both branches
-        let commit1 = self.get_branch_commit(branch1)?;
-        let commit2 = self.get_branch_commit(branch2)?;
-
-        // For now, use a simple approach: find common ancestor by walking history
-        // This is a simplified implementation - a real merge-base algorithm would be more sophisticated
-        let mut visited1 = std::collections::HashSet::new();
-        let mut queue1 = std::collections::VecDeque::new();
-        queue1.push_back(commit1);
-
-        // Collect all ancestors of branch1
-        while let Some(commit_id) = queue1.pop_front() {
-            if visited1.contains(&commit_id) {
-                continue;
-            }
-            visited1.insert(commit_id);
-
-            // Add parents
-            if let Ok(parents) = self.get_commit_parents(&commit_id) {
-                for parent in parents {
-                    if !visited1.contains(&parent) {
-                        queue1.push_back(parent);
-                    }
-                }
-            }
-        }
-
-        // Walk branch2 and find first common ancestor
-        let mut visited2 = std::collections::HashSet::new();
-        let mut queue2 = std::collections::VecDeque::new();
-        queue2.push_back(commit2);
-
-        while let Some(commit_id) = queue2.pop_front() {
-            if visited2.contains(&commit_id) {
-                continue;
-            }
-            visited2.insert(commit_id);
-
-            // Check if this commit is in branch1's ancestors
-            if visited1.contains(&commit_id) {
-                return Ok(commit_id);
-            }
-
-            // Add parents
-            if let Ok(parents) = self.get_commit_parents(&commit_id) {
-                for parent in parents {
-                    if !visited2.contains(&parent) {
-                        queue2.push_back(parent);
-                    }
-                }
-            }
-        }
-
-        Err(GitKvError::GitObjectError(
-            "No common ancestor found".to_string(),
-        ))
-    }
-
-    /// Get commit ID for a branch
-    fn get_branch_commit(&self, branch: &str) -> Result<gix::ObjectId, GitKvError> {
-        let branch_ref = format!("refs/heads/{branch}");
-        match self.git_repo.refs.find(&branch_ref) {
-            Ok(reference) => match reference.target.try_id() {
-                Some(commit_id) => Ok(commit_id.to_owned()),
-                None => Err(GitKvError::GitObjectError(format!(
-                    "Branch {branch} does not point to a commit"
-                ))),
-            },
-            Err(_) => Err(GitKvError::BranchNotFound(branch.to_string())),
-        }
-    }
-
-    /// Get parents of a commit
-    fn get_commit_parents(
-        &self,
-        commit_id: &gix::ObjectId,
-    ) -> Result<Vec<gix::ObjectId>, GitKvError> {
-        let mut buffer = Vec::new();
-        let commit_obj = self
-            .git_repo
-            .objects
-            .find(commit_id, &mut buffer)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to find commit: {e}")))?;
-
-        let parents = match commit_obj.decode() {
-            Ok(gix::objs::ObjectRef::Commit(commit)) => commit.parents().collect(),
-            _ => {
-                return Err(GitKvError::GitObjectError(
-                    "Object is not a commit".to_string(),
-                ))
-            }
-        };
-        Ok(parents)
-    }
-
-    /// Create a merge commit
-    fn create_merge_commit(
-        &mut self,
-        message: &str,
-        source_branch: &str,
-    ) -> Result<gix::ObjectId, GitKvError> {
-        // Get the source branch commit as second parent
-        let _source_commit = self.get_branch_commit(source_branch)?;
-
-        // Save current tree state
-        self.tree.persist_root();
-        self.tree
-            .save_config()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to save config: {e}")))?;
-
-        // Save tree config to git
-        self.save_tree_config_to_git_internal()?;
-
-        // Create git tree and commit (similar to regular commit but with two parents)
-        // Use work_dir() for worktree/submodule compatibility
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".into()))?;
-        let git_root = self
-            .git_repo
-            .work_dir()
-            .map(|p| p.to_path_buf())
-            .or_else(|| Self::find_git_root(dataset_dir))
-            .ok_or_else(|| GitKvError::GitObjectError("Could not find git root".into()))?;
-
-        // Stage all files
-        let add_output = std::process::Command::new("git")
-            .args(["add", "."])
-            .current_dir(&git_root)
-            .output()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to run git add: {e}")))?;
-
-        if !add_output.status.success() {
-            return Err(GitKvError::GitObjectError(format!(
-                "Git add failed: {}",
-                String::from_utf8_lossy(&add_output.stderr)
-            )));
-        }
-
-        // Create merge commit with two parents
-        let _current_commit = self.git_repo.head_id().map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to get current commit: {e}"))
-        })?;
-
-        let commit_output = std::process::Command::new("git")
-            .args(["commit", "-m", message, "--allow-empty"])
-            .current_dir(&git_root)
-            .env("GIT_AUTHOR_NAME", "ProllyTree")
-            .env("GIT_AUTHOR_EMAIL", "prollytree@example.com")
-            .env("GIT_COMMITTER_NAME", "ProllyTree")
-            .env("GIT_COMMITTER_EMAIL", "prollytree@example.com")
-            .output()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to run git commit: {e}")))?;
-
-        if !commit_output.status.success() {
-            return Err(GitKvError::GitObjectError(format!(
-                "Git commit failed: {}",
-                String::from_utf8_lossy(&commit_output.stderr)
-            )));
-        }
-
-        // Get the new commit ID
-        let new_commit = self
-            .git_repo
-            .head_id()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get new commit: {e}")))?;
-
-        Ok(new_commit.into())
-    }
-
-    /// Reload the tree state from the current HEAD commit
-    fn reload_tree_from_head(&mut self) -> Result<(), GitKvError> {
-        // Get the current HEAD commit
-        let head = self
-            .git_repo
-            .head()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get HEAD: {e}")))?;
-
-        let head_commit_id = head.id().ok_or_else(|| {
-            GitKvError::GitObjectError("HEAD does not point to a commit".to_string())
-        })?;
-
-        // Convert gix::Id to gix::ObjectId
-        let head_object_id = head_commit_id.detach();
-
-        // Load all key-value pairs from the HEAD commit
-        let keys_at_head = self.collect_keys_at_commit(&head_object_id)?;
-
-        // Clear the current tree and rebuild it with the data from HEAD
-        // Important: Create tree with config that already has the correct root hash to avoid triggering persist_root
-        let mut config = self.tree.config.clone();
-
-        // Get the config from the commit to get the correct root hash
-        if let Ok(commit_config) = self.read_tree_config_from_commit(&head_object_id) {
-            config.root_hash = commit_config.root_hash;
-        }
-
-        self.tree = ProllyTree::new(self.tree.storage.clone(), config);
-
-        // Insert all the key-value pairs from the HEAD commit
-        // Note: These insertions may still trigger auto-save in the tree
-        for (key, value) in keys_at_head {
-            self.tree.insert(key, value);
-        }
-
-        // Note: We don't save_config() or persist_root() here because this is a read operation.
-        // The config files should only be saved during write operations (commit).
-        // Saving here would overwrite the existing files with potentially stale data.
-
-        Ok(())
-    }
-
-    /// Initialize a new versioned KV store with Git storage (default)
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-
-        // Safety check: prevent initializing at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot initialize git-prolly in git root directory. \
-                Please use a subdirectory to create a dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // For GitVersionedKvStore, use the user-provided path as the dataset directory
-        // This allows config files to be versioned in git commits
-        let dataset_dir = path.to_path_buf();
-        std::fs::create_dir_all(&dataset_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create dataset directory: {e}"))
-        })?;
-
-        // Check if the store is already initialized by looking for config files
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        let mappings_path = dataset_dir.join("prolly_hash_mappings");
-
-        if config_path.exists() || mappings_path.exists() {
-            // Store already exists, use open instead to load existing configuration
-            return Self::open(path);
-        }
-
-        // Open the existing git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Create GitNodeStorage with user-provided path for versioned files
-        let storage = GitNodeStorage::new(git_repo.clone(), dataset_dir.clone())?;
-
-        // Create ProllyTree with default config
-        let config: TreeConfig<N> = TreeConfig::default();
-        let tree = ProllyTree::new(storage, config);
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch: "main".to_string(),
-            storage_backend: StorageBackend::Git,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Save initial configuration
-        let _ = store.tree.save_config();
-
-        // Create initial commit (which will include prolly metadata files)
-        store.commit("Initial commit")?;
-
-        Ok(store)
-    }
-
-    /// Open an existing versioned KV store with Git storage (default)
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-
-        // Safety check: prevent opening at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot open git-prolly in git root directory. \
-                Please use a subdirectory for your dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // For GitVersionedKvStore, use the user-provided path as the dataset directory
-        // This allows config files to be versioned in git commits
-        let dataset_dir = path.to_path_buf();
-
-        // Open existing Git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Create GitNodeStorage with user-provided path for versioned files
-        let storage = GitNodeStorage::new(git_repo.clone(), dataset_dir.clone())?;
-
-        // Load tree configuration from dataset directory
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        if !config_path.exists() {
-            return Err(GitKvError::GitObjectError(
-                "Config file not found. The store may not be initialized. \
-                Call init() to create a new store."
-                    .to_string(),
-            ));
-        }
-        let config_data = std::fs::read_to_string(&config_path)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to read config file: {e}")))?;
-        let config: TreeConfig<N> = serde_json::from_str(&config_data)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to parse config file: {e}")))?;
-
-        // Try to load existing tree from storage
-        let tree = if let Some(existing_tree) =
-            ProllyTree::load_from_storage(storage.clone(), config.clone())
-        {
-            existing_tree
-        } else if config.root_hash.is_some() {
-            // We have a saved root hash but failed to load the tree
-            // This could be due to missing hash mappings or git objects
-            // For read-only operations, we should try to work with what we have
-            // rather than creating a new empty tree that would overwrite the config
-            eprintln!("Warning: Failed to load tree from saved root hash. This may indicate missing git objects or corrupted hash mappings.");
-            eprintln!("Attempting to create tree with saved config to avoid data loss...");
-            ProllyTree::new(storage, config)
-        } else {
-            // No saved root hash - this is a genuinely new/empty tree
-            ProllyTree::new(storage, config)
-        };
-
-        // Get current branch
-        let current_branch = git_repo
-            .head_ref()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get head ref: {e}")))?
-            .map(|r| r.name().shorten().to_string())
-            .unwrap_or_else(|| "main".to_string());
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch,
-            storage_backend: StorageBackend::Git,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Load staging area from file if it exists
-        store.load_staging_area()?;
-
-        // Note: We intentionally do NOT call reload_tree_from_head() here
-        // because git-prolly commands should read from the current directory's
-        // prolly_config_tree_config and mapping files, not from git HEAD.
-        // The tree was already loaded from local storage by load_from_storage() above.
-
-        Ok(store)
-    }
-
-    /// Get a reference to the underlying ProllyTree
-    pub fn tree(&self) -> &ProllyTree<N, GitNodeStorage<N>> {
-        &self.tree
-    }
-
-    /// Get a mutable reference to the underlying ProllyTree
-    pub fn tree_mut(&mut self) -> &mut ProllyTree<N, GitNodeStorage<N>> {
-        &mut self.tree
-    }
-
-    /// Collect all key-value pairs from the tree at a specific commit
-    fn collect_keys_at_commit(
-        &self,
-        commit_id: &gix::ObjectId,
-    ) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Get the commit object
-        let mut buffer = Vec::new();
-        let commit = self
-            .git_repo
-            .objects
-            .find(commit_id, &mut buffer)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to find commit: {e}")))?;
-
-        let commit_ref = commit
-            .decode()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to decode commit: {e}")))?
-            .into_commit()
-            .ok_or_else(|| GitKvError::GitObjectError("Object is not a commit".to_string()))?;
-
-        // Get the tree object from the commit
-        let tree_id = commit_ref.tree();
-
-        // Try to load the prolly tree configuration from the tree
-        // Since files are stored relative to git root, we need the relative path from git root to dataset
-        // Use work_dir() for worktree root (handles worktrees/submodules correctly)
-        let dataset_dir = self.tree.storage.dataset_dir();
-        let git_root = self
-            .git_repo
-            .work_dir()
-            .map(|p| p.to_path_buf())
-            .or_else(|| Self::find_git_root(dataset_dir))
-            .ok_or_else(|| GitKvError::GitObjectError("Could not find git root".to_string()))?;
-
-        let dataset_relative_path = dataset_dir
-            .strip_prefix(&git_root)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get relative path: {e}")))?;
-
-        // Build paths using '/' separators for git tree lookups (platform-independent)
-        let relative_path_str = dataset_relative_path
-            .components()
-            .map(|c| c.as_os_str().to_string_lossy())
-            .collect::<Vec<_>>()
-            .join("/");
-
-        let config_path = format!("{}/prolly_config_tree_config", relative_path_str);
-        let mapping_path = format!("{}/prolly_hash_mappings", relative_path_str);
-
-        let config_result = self.read_file_from_tree(&tree_id, &config_path);
-        let mapping_result = self.read_file_from_tree(&tree_id, &mapping_path);
-
-        // If files are not found, this might be an initial empty commit, return empty
-        if config_result.is_err() || mapping_result.is_err() {
-            return Ok(HashMap::new());
-        }
-
-        let config_data = config_result?;
-        let config: TreeConfig<N> = serde_json::from_slice(&config_data).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to deserialize config: {e}"))
-        })?;
-
-        // Load the hash mappings from the tree as string format and parse
-        let mapping_data = mapping_result?;
-        let mapping_str = String::from_utf8(mapping_data)
-            .map_err(|e| GitKvError::GitObjectError(format!("Invalid UTF-8 in mappings: {e}")))?;
-
-        // Check if this is a simple key-value mapping (for InMemory storage)
-        // or a hash mapping (for Git storage)
-        let mut key_values = HashMap::new();
-        let mut hash_mappings = HashMap::new();
-        let mut is_simple_mapping = false;
-
-        for line in mapping_str.lines() {
-            if let Some((prefix, rest)) = line.split_once(':') {
-                if prefix == "key" {
-                    // This is a simple key-value mapping from InMemory storage
-                    is_simple_mapping = true;
-                    if let Some((key_hex, value_hex)) = rest.split_once(':') {
-                        if let (Ok(key), Ok(value)) = (hex::decode(key_hex), hex::decode(value_hex))
-                        {
-                            key_values.insert(key, value);
-                        }
-                    }
-                } else {
-                    // This is a hash mapping from Git storage
-                    let hash_hex = prefix;
-                    let object_hex = rest;
-
-                    if hash_hex.len() == N * 2 {
-                        let mut hash_bytes = Vec::new();
-                        for i in 0..N {
-                            if let Ok(byte) = u8::from_str_radix(&hash_hex[i * 2..i * 2 + 2], 16) {
-                                hash_bytes.push(byte);
-                            } else {
-                                break;
-                            }
-                        }
-
-                        if hash_bytes.len() == N {
-                            if let Ok(object_id) = gix::ObjectId::from_hex(object_hex.as_bytes()) {
-                                let hash = ValueDigest::raw_hash(&hash_bytes);
-                                hash_mappings.insert(hash, object_id);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        if is_simple_mapping {
-            // For InMemory storage, return the directly stored key-value pairs
-            return Ok(key_values);
-        }
-
-        // For Git storage, reconstruct the tree from hash mappings
-        if hash_mappings.is_empty() {
-            return Ok(HashMap::new());
-        }
-
-        // Create a temporary storage with the loaded mappings
-        let temp_storage = GitNodeStorage::with_mappings(
-            self.git_repo.clone(),
-            self.tree.storage.dataset_dir().to_path_buf(),
-            hash_mappings,
-        )?;
-
-        // Load the tree with the config
-        let tree = ProllyTree::load_from_storage(temp_storage, config).ok_or_else(|| {
-            GitKvError::GitObjectError("Failed to load tree from storage".to_string())
-        })?;
-
-        // Collect all key-value pairs
-        let mut result_key_values = HashMap::new();
-        for key in tree.collect_keys() {
-            if let Some(node) = tree.find(&key) {
-                // Find the value in the node
-                if let Some(index) = node.keys.iter().position(|k| k == &key) {
-                    result_key_values.insert(key, node.values[index].clone());
-                }
-            }
-        }
-
-        Ok(result_key_values)
-    }
-}
-
-// Implement HistoricalAccess for GitNodeStorage
-impl<const N: usize> HistoricalAccess<N> for VersionedKvStore<N, GitNodeStorage<N>> {
-    fn get_keys_at_ref(&self, reference: &str) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        let commit_id = self.resolve_commit(reference)?;
-        self.collect_keys_at_commit(&commit_id)
-    }
-}
-
-// Implement HistoricalCommitAccess for GitNodeStorage
-impl<const N: usize> HistoricalCommitAccess<N> for VersionedKvStore<N, GitNodeStorage<N>> {
-    fn get_commits_for_key(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        let mut commit_history = self.get_commit_history()?;
-
-        // Reverse to process in chronological order (oldest first)
-        commit_history.reverse();
-
-        let mut commits_with_key_changes = Vec::new();
-        let mut previous_value: Option<Vec<u8>> = None; // None = key not present, Some(val) = key present with value
-
-        for commit in commit_history {
-            // Get the key value at this commit
-            let current_value = self.collect_keys_at_commit(&commit.id)?.get(key).cloned();
-
-            // Check if the value changed from the previous commit
-            let value_changed = previous_value != current_value;
-
-            if value_changed {
-                commits_with_key_changes.push(commit);
-            }
-
-            previous_value = current_value;
-        }
-
-        // Reverse back to newest first for the final result
-        commits_with_key_changes.reverse();
-
-        Ok(commits_with_key_changes)
-    }
-
-    fn get_commit_history(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        // Reuse the existing log method
-        self.log()
-    }
-}
-
-impl<const N: usize> VersionedKvStore<N, InMemoryNodeStorage<N>> {
-    /// Initialize a new versioned KV store with InMemory storage
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-
-        // Safety check: prevent initializing at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot initialize in-memory store in git root directory. \
-                Please use a subdirectory to create a dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // Open the existing git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Create dataset directory for config files
-        let dataset_dir = path.to_path_buf();
-        std::fs::create_dir_all(&dataset_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create dataset directory: {e}"))
-        })?;
-
-        // Create InMemoryNodeStorage
-        let storage = InMemoryNodeStorage::<N>::new();
-
-        // Create ProllyTree with default config
-        let config: TreeConfig<N> = TreeConfig::default();
-        let tree = ProllyTree::new(storage, config);
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch: "main".to_string(),
-            storage_backend: StorageBackend::InMemory,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Create initial commit
-        store.commit("Initial commit")?;
-
-        Ok(store)
-    }
-
-    /// Open an existing versioned KV store with InMemory storage
-    /// Note: InMemory storage is volatile, so this creates a new empty store
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        // For InMemory storage, "opening" is the same as initializing
-        // since data is not persistent
-        Self::init(path)
-    }
-}
-
-// Implement HistoricalAccess for InMemoryNodeStorage
-impl<const N: usize> HistoricalAccess<N> for VersionedKvStore<N, InMemoryNodeStorage<N>> {
-    fn get_keys_at_ref(&self, reference: &str) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Resolve the reference to a commit ID
-        let commit_id = self.resolve_commit(reference)?;
-
-        // Get the tree config from the commit to extract root hash
-        let tree_config = self.read_tree_config_from_commit(&commit_id)?;
-
-        // Reconstruct the tree state from storage using the root hash
-        self.collect_keys_from_config(&tree_config)
-    }
-}
-
-// Implement HistoricalCommitAccess for InMemoryNodeStorage
-impl<const N: usize> HistoricalCommitAccess<N> for VersionedKvStore<N, InMemoryNodeStorage<N>> {
-    fn get_commits_for_key(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commits_for_key_generic(key)
-    }
-
-    fn get_commit_history(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commit_history_generic()
-    }
-}
-
-impl<const N: usize> VersionedKvStore<N, FileNodeStorage<N>> {
-    /// Initialize a new versioned KV store with File storage
-    ///
-    /// Nodes are stored in `.git/prolly/nodes/files/` (shared, content-addressed).
-    /// Config is stored in the dataset directory (committed to git for history).
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-
-        // Safety check: prevent initializing at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot initialize file store in git root directory. \
-                Please use a subdirectory to create a dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // Create prolly directory inside .git for node storage
-        let prolly_dir = Self::ensure_prolly_dir(&git_root)?;
-
-        // Create dataset directory for config files (will be committed to git)
-        let dataset_dir = path.to_path_buf();
-        std::fs::create_dir_all(&dataset_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create dataset directory: {e}"))
-        })?;
-
-        // Open the existing git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Create FileNodeStorage inside .git/prolly/nodes/files/ (shared across datasets)
-        let file_storage_path = prolly_dir.join("nodes").join("files");
-        std::fs::create_dir_all(&file_storage_path).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create file storage directory: {e}"))
-        })?;
-        let storage = FileNodeStorage::<N>::new(file_storage_path);
-
-        // Create ProllyTree with default config
-        let config: TreeConfig<N> = TreeConfig::default();
-        let tree = ProllyTree::new(storage, config);
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch: "main".to_string(),
-            storage_backend: StorageBackend::File,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Create initial commit (which will save config to dataset_dir)
-        store.commit("Initial commit")?;
-
-        Ok(store)
-    }
-
-    /// Open an existing versioned KV store with File storage
-    ///
-    /// Nodes are loaded from `.git/prolly/nodes/files/` (shared, content-addressed).
-    /// Config is loaded from the dataset directory.
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-        let dataset_dir = path.to_path_buf();
-
-        // Safety check: prevent opening at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot open file store in git root directory. \
-                Please use a subdirectory for your dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Check if the dataset directory exists
-        if !dataset_dir.exists() {
-            return Err(GitKvError::GitObjectError(
-                "Dataset directory not found. Call init() first to create the store.".to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // Get prolly directory inside .git
-        let prolly_dir = Self::get_prolly_dir(&git_root);
-
-        // Open existing Git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Open FileNodeStorage inside .git/prolly/nodes/files/
-        let file_storage_path = prolly_dir.join("nodes").join("files");
-
-        // Check if the file storage directory exists - if not, the store hasn't been initialized
-        if !file_storage_path.exists() {
-            return Err(GitKvError::GitObjectError(
-                "File store not initialized. Call init() first to create the store.".to_string(),
-            ));
-        }
-
-        let storage = FileNodeStorage::<N>::new(file_storage_path.clone());
-
-        // Load tree configuration from dataset directory (not from storage)
-        // Config file must exist for open() - use init() to create new stores
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        if !config_path.exists() {
-            return Err(GitKvError::GitObjectError(
-                "Config file not found. The store may not be initialized. \
-                Call init() to create a new store."
-                    .to_string(),
-            ));
-        }
-        let config_data = std::fs::read_to_string(&config_path)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to read config file: {e}")))?;
-        let config: TreeConfig<N> = serde_json::from_str(&config_data)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to parse config file: {e}")))?;
-
-        // Try to load existing tree from storage using the config's root hash
-        let tree =
-            if let Some(existing_tree) = ProllyTree::load_from_storage(storage, config.clone()) {
-                existing_tree
-            } else {
-                // Create new storage instance since the original was consumed
-                let new_storage = FileNodeStorage::<N>::new(file_storage_path);
-                ProllyTree::new(new_storage, config)
-            };
-
-        // Get current branch
-        let current_branch = git_repo
-            .head_ref()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get head ref: {e}")))?
-            .map(|r| r.name().shorten().to_string())
-            .unwrap_or_else(|| "main".to_string());
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch,
-            storage_backend: StorageBackend::File,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Load staging area from file if it exists
-        store.load_staging_area()?;
-
-        Ok(store)
-    }
-}
-
-// Implement HistoricalAccess for FileNodeStorage
-impl<const N: usize> HistoricalAccess<N> for VersionedKvStore<N, FileNodeStorage<N>> {
-    fn get_keys_at_ref(&self, reference: &str) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Resolve the reference to a commit ID
-        let commit_id = self.resolve_commit(reference)?;
-
-        // Get the tree config from the commit to extract root hash
-        let tree_config = self.read_tree_config_from_commit(&commit_id)?;
-
-        // Reconstruct the tree state from storage using the root hash
-        self.collect_keys_from_config(&tree_config)
-    }
-}
-
-// Implement HistoricalCommitAccess for FileNodeStorage
-impl<const N: usize> HistoricalCommitAccess<N> for VersionedKvStore<N, FileNodeStorage<N>> {
-    fn get_commits_for_key(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commits_for_key_generic(key)
-    }
-
-    fn get_commit_history(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commit_history_generic()
-    }
-}
-
-#[cfg(feature = "rocksdb_storage")]
-impl<const N: usize> VersionedKvStore<N, RocksDBNodeStorage<N>> {
-    /// Initialize a new versioned KV store with RocksDB storage
-    ///
-    /// Nodes are stored in `.git/prolly/nodes/rocksdb/` (shared, content-addressed).
-    /// Config is stored in the dataset directory (committed to git for history).
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-
-        // Safety check: prevent initializing at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot initialize RocksDB store in git root directory. \
-                Please use a subdirectory to create a dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // Create prolly directory inside .git for node storage
-        let prolly_dir = Self::ensure_prolly_dir(&git_root)?;
-
-        // Create dataset directory for config files (will be committed to git)
-        let dataset_dir = path.to_path_buf();
-        std::fs::create_dir_all(&dataset_dir).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create dataset directory: {e}"))
-        })?;
-
-        // Open the existing git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Create RocksDBNodeStorage inside .git/prolly/nodes/rocksdb/ (shared across datasets)
-        let rocksdb_path = prolly_dir.join("nodes").join("rocksdb");
-        std::fs::create_dir_all(&rocksdb_path).map_err(|e| {
-            GitKvError::GitObjectError(format!("Failed to create RocksDB directory: {e}"))
-        })?;
-        let storage = RocksDBNodeStorage::<N>::new(rocksdb_path)
-            .map_err(|e| GitKvError::GitObjectError(format!("RocksDB creation failed: {e}")))?;
-
-        // Create ProllyTree with default config
-        let config: TreeConfig<N> = TreeConfig::default();
-        let tree = ProllyTree::new(storage, config);
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch: "main".to_string(),
-            storage_backend: StorageBackend::RocksDB,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Create initial commit (which will save config to dataset_dir)
-        store.commit("Initial commit")?;
-
-        Ok(store)
-    }
-
-    /// Open an existing versioned KV store with RocksDB storage
-    ///
-    /// Nodes are loaded from `.git/prolly/nodes/rocksdb/` (shared, content-addressed).
-    /// Config is loaded from the dataset directory.
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let path = path.as_ref();
-        let dataset_dir = path.to_path_buf();
-
-        // Safety check: prevent opening at git root to avoid `git add -A .` staging all files
-        if Self::is_in_git_root(path)? {
-            return Err(GitKvError::GitObjectError(
-                "Cannot open RocksDB store in git root directory. \
-                Please use a subdirectory for your dataset, or the commit operation \
-                may accidentally stage all files in the repository."
-                    .to_string(),
-            ));
-        }
-
-        // Check if the dataset directory exists
-        if !dataset_dir.exists() {
-            return Err(GitKvError::GitObjectError(
-                "Dataset directory not found. Call init() first to create the store.".to_string(),
-            ));
-        }
-
-        // Find the git repository
-        let git_root = Self::find_git_root(path).ok_or_else(|| {
-            GitKvError::GitObjectError(
-                "Not inside a git repository. Please run from within a git repository.".to_string(),
-            )
-        })?;
-
-        // Get prolly directory inside .git
-        let prolly_dir = Self::get_prolly_dir(&git_root);
-
-        // Open existing Git repository
-        let git_repo = gix::open(&git_root).map_err(|e| GitKvError::GitOpenError(Box::new(e)))?;
-
-        // Open RocksDBNodeStorage inside .git/prolly/nodes/rocksdb/
-        let rocksdb_path = prolly_dir.join("nodes").join("rocksdb");
-
-        // Check if the RocksDB directory exists - if not, the store hasn't been initialized
-        if !rocksdb_path.exists() {
-            return Err(GitKvError::GitObjectError(
-                "RocksDB store not initialized. Call init() first to create the store.".to_string(),
-            ));
-        }
-
-        let storage = RocksDBNodeStorage::<N>::new(rocksdb_path)
-            .map_err(|e| GitKvError::GitObjectError(format!("RocksDB creation failed: {e}")))?;
-
-        // Load tree configuration from dataset directory (not from storage)
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        if !config_path.exists() {
-            return Err(GitKvError::GitObjectError(
-                "Config file not found. The store may not be initialized. \
-                Call init() to create a new store."
-                    .to_string(),
-            ));
-        }
-        let config_data = std::fs::read_to_string(&config_path)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to read config file: {e}")))?;
-        let config: TreeConfig<N> = serde_json::from_str(&config_data)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to parse config file: {e}")))?;
-
-        // Try to load existing tree from storage using the config's root hash
-        let tree = ProllyTree::load_from_storage(storage.clone(), config.clone())
-            .unwrap_or_else(|| ProllyTree::new(storage, config));
-
-        // Get current branch
-        let current_branch = git_repo
-            .head_ref()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get head ref: {e}")))?
-            .map(|r| r.name().shorten().to_string())
-            .unwrap_or_else(|| "main".to_string());
-
-        let mut store = VersionedKvStore {
-            tree,
-            git_repo,
-            staging_area: HashMap::new(),
-            current_branch,
-            storage_backend: StorageBackend::RocksDB,
-            dataset_dir: Some(dataset_dir),
-        };
-
-        // Load staging area from file if it exists
-        store.load_staging_area()?;
-
-        Ok(store)
-    }
-}
-
-// Implement HistoricalAccess for RocksDBNodeStorage
-#[cfg(feature = "rocksdb_storage")]
-impl<const N: usize> HistoricalAccess<N> for VersionedKvStore<N, RocksDBNodeStorage<N>> {
-    fn get_keys_at_ref(&self, reference: &str) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Resolve the reference to a commit ID
-        let commit_id = self.resolve_commit(reference)?;
-
-        // Get the tree config from the commit to extract root hash
-        let tree_config = self.read_tree_config_from_commit(&commit_id)?;
-
-        // Reconstruct the tree state from storage using the root hash
-        self.collect_keys_from_config(&tree_config)
-    }
-}
-
-// Implement HistoricalCommitAccess for RocksDBNodeStorage
-#[cfg(feature = "rocksdb_storage")]
-impl<const N: usize> HistoricalCommitAccess<N> for VersionedKvStore<N, RocksDBNodeStorage<N>> {
-    fn get_commits_for_key(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commits_for_key_generic(key)
-    }
-
-    fn get_commit_history(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        self.get_commit_history_generic()
-    }
-}
-
-// Generic implementations for all storage types that support historical access
-impl<const N: usize, S: NodeStorage<N>> VersionedKvStore<N, S>
-where
-    Self: TreeConfigSaver<N>,
-{
-    /// Get the current storage backend type
-    pub fn storage_backend(&self) -> &StorageBackend {
-        &self.storage_backend
-    }
-
-    /// Resolve a reference (branch name, commit SHA, etc.) to a commit ID
-    /// This is used by all storage types for historical access
-    fn resolve_commit(&self, reference: &str) -> Result<gix::ObjectId, GitKvError> {
-        // Try to resolve as a branch first
-        if let Ok(mut branch_ref) = self
-            .git_repo
-            .find_reference(&format!("refs/heads/{reference}"))
-        {
-            // Try to peel the reference to get the commit ID
-            if let Ok(peeled) = branch_ref.peel_to_id_in_place() {
-                return Ok(peeled.detach());
-            }
-        }
-
-        // Try to resolve as a commit SHA
-        if let Ok(commit_id) = gix::ObjectId::from_hex(reference.as_bytes()) {
-            // Verify the commit exists by trying to find it
-            let mut buffer = Vec::new();
-            if self.git_repo.objects.find(&commit_id, &mut buffer).is_ok() {
-                return Ok(commit_id);
-            }
-        }
-
-        // Try other reference formats (tags, etc.)
-        if let Ok(mut reference) = self.git_repo.find_reference(reference) {
-            // Try to peel the reference to get the commit ID
-            if let Ok(peeled) = reference.peel_to_id_in_place() {
-                return Ok(peeled.detach());
-            }
-        }
-
-        Err(GitKvError::InvalidCommit(format!(
-            "Reference '{reference}' not found"
-        )))
-    }
-
-    /// Get the config file path relative to git root
-    /// All backends now store dataset_dir, so this is consistent across all storage types
-    fn get_config_file_path_relative_to_git_root(&self) -> Result<String, GitKvError> {
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".to_string()))?;
-
-        // Use work_dir() for worktree root (handles worktrees/submodules correctly),
-        // falling back to find_git_root(dataset_dir) if work_dir is not available
-        let git_root = self
-            .git_repo
-            .work_dir()
-            .map(|p| p.to_path_buf())
-            .or_else(|| Self::find_git_root(dataset_dir))
-            .ok_or_else(|| GitKvError::GitObjectError("Could not find git root".to_string()))?;
-
-        // Calculate relative path from git root to dataset dir
-        let relative_path = dataset_dir.strip_prefix(&git_root).map_err(|_| {
-            GitKvError::GitObjectError("Dataset directory is not inside git repository".to_string())
-        })?;
-
-        // Construct the file path and use '/' separators for git tree paths
-        // (git uses forward slashes regardless of platform)
-        let config_path = relative_path.join("prolly_config_tree_config");
-        let path_str = config_path
-            .components()
-            .map(|c| c.as_os_str().to_string_lossy())
-            .collect::<Vec<_>>()
-            .join("/");
-        Ok(path_str)
-    }
-
-    /// Read the tree config from a specific commit
-    /// This gets the prolly_config_tree_config file from the commit to extract root hash
-    fn read_tree_config_from_commit(
-        &self,
-        commit_id: &gix::ObjectId,
-    ) -> Result<TreeConfig<N>, GitKvError> {
-        // Get the config file path relative to git root
-        let config_path = self.get_config_file_path_relative_to_git_root()?;
-
-        // Get the commit object
-        let mut commit_buffer = Vec::new();
-        let commit_obj = self
-            .git_repo
-            .objects
-            .find(commit_id, &mut commit_buffer)
-            .map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to find commit {commit_id}: {e}"))
-            })?;
-
-        let commit = match commit_obj.kind {
-            gix::object::Kind::Commit => gix::objs::CommitRef::from_bytes(commit_obj.data)
-                .map_err(|e| GitKvError::GitObjectError(format!("Failed to parse commit: {e}")))?,
-            _ => {
-                return Err(GitKvError::InvalidCommit(format!(
-                    "{commit_id} is not a commit"
-                )))
-            }
-        };
-
-        // Get the tree object
-        let tree_id = commit.tree();
-
-        // Try to read the config file, with fallback to current config if not found
-        match self.read_file_from_tree(&tree_id, &config_path) {
-            Ok(config_data) => {
-                // Parse the config
-                let tree_config: TreeConfig<N> =
-                    serde_json::from_slice(&config_data).map_err(|e| {
-                        GitKvError::GitObjectError(format!("Failed to parse tree config: {e}"))
-                    })?;
-                Ok(tree_config)
-            }
-            Err(_) => {
-                // If config file is not found in commit, create a default config
-                // This can happen for commits that don't have prolly config saved
-                // or for initial commits before the config system was in place
-                eprintln!("Warning: prolly_config_tree_config not found in commit {commit_id}, using default config");
-                Ok(TreeConfig::default())
-            }
-        }
-    }
-
-    /// Read a file from a git tree (helper for all storage types)
-    fn read_file_from_tree(
-        &self,
-        tree_id: &gix::ObjectId,
-        file_path: &str,
-    ) -> Result<Vec<u8>, GitKvError> {
-        let mut tree_buffer = Vec::new();
-        let tree_obj = self
-            .git_repo
-            .objects
-            .find(tree_id, &mut tree_buffer)
-            .map_err(|e| {
-                GitKvError::GitObjectError(format!("Failed to find tree {tree_id}: {e}"))
-            })?;
-
-        let tree = match tree_obj.kind {
-            gix::object::Kind::Tree => gix::objs::TreeRef::from_bytes(tree_obj.data)
-                .map_err(|e| GitKvError::GitObjectError(format!("Failed to parse tree: {e}")))?,
-            _ => {
-                return Err(GitKvError::GitObjectError(format!(
-                    "{tree_id} is not a tree"
-                )))
-            }
-        };
-
-        // Handle file paths with subdirectories
-        let path_parts: Vec<&str> = file_path.split('/').collect();
-        self.read_file_from_tree_recursive(&tree, &path_parts, 0)
-    }
-
-    /// Recursively read a file from a tree, handling subdirectories
-    fn read_file_from_tree_recursive(
-        &self,
-        tree: &gix::objs::TreeRef,
-        path_parts: &[&str],
-        part_index: usize,
-    ) -> Result<Vec<u8>, GitKvError> {
-        if part_index >= path_parts.len() {
-            return Err(GitKvError::GitObjectError(
-                "Path traversal error".to_string(),
-            ));
-        }
-
-        let current_part = path_parts[part_index];
-
-        // Search for the current path part in the tree
-        for entry in &tree.entries {
-            if entry.filename == current_part.as_bytes() {
-                if part_index == path_parts.len() - 1 {
-                    // This is the final part (the file), read its content
-                    let mut file_buffer = Vec::new();
-                    let file_obj = self
-                        .git_repo
-                        .objects
-                        .find(entry.oid, &mut file_buffer)
-                        .map_err(|e| {
-                            GitKvError::GitObjectError(format!("Failed to find file object: {e}"))
-                        })?;
-
-                    match file_obj.kind {
-                        gix::object::Kind::Blob => return Ok(file_obj.data.to_vec()),
-                        _ => {
-                            return Err(GitKvError::GitObjectError(
-                                "File is not a blob".to_string(),
-                            ))
-                        }
-                    }
-                } else {
-                    // This is a directory, recurse into it
-                    let mut subtree_buffer = Vec::new();
-                    let subtree_obj = self
-                        .git_repo
-                        .objects
-                        .find(entry.oid, &mut subtree_buffer)
-                        .map_err(|e| {
-                            GitKvError::GitObjectError(format!(
-                                "Failed to find subtree object: {e}"
-                            ))
-                        })?;
-
-                    match subtree_obj.kind {
-                        gix::object::Kind::Tree => {
-                            let subtree = gix::objs::TreeRef::from_bytes(subtree_obj.data)
-                                .map_err(|e| {
-                                    GitKvError::GitObjectError(format!(
-                                        "Failed to parse subtree: {e}"
-                                    ))
-                                })?;
-                            return self.read_file_from_tree_recursive(
-                                &subtree,
-                                path_parts,
-                                part_index + 1,
-                            );
-                        }
-                        _ => {
-                            return Err(GitKvError::GitObjectError(
-                                "Expected directory but found file".to_string(),
-                            ))
-                        }
-                    }
-                }
-            }
-        }
-
-        Err(GitKvError::GitObjectError(format!(
-            "Path component '{current_part}' not found in tree"
-        )))
-    }
-
-    /// Collect all key-value pairs from storage using a tree config (with root hash)
-    /// This reconstructs the tree state for non-git storage types
-    fn collect_keys_from_config(
-        &self,
-        tree_config: &TreeConfig<N>,
-    ) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Get the root hash from the config
-        let root_hash = match tree_config.root_hash.as_ref() {
-            Some(hash) => hash,
-            None => {
-                // If no root hash in config, return empty result
-                // This can happen for initial commits or when config wasn't properly saved
-                eprintln!("Warning: No root hash in tree config, returning empty key set");
-                return Ok(HashMap::new());
-            }
-        };
-
-        // Reconstruct the tree from storage using the root hash
-        let root_node = match self.tree.storage.get_node_by_hash(root_hash) {
-            Some(node) => node,
-            None => {
-                // Root node not found in storage, return empty result
-                // This can happen if the historical state is not available in current storage
-                eprintln!("Warning: Root node not found in storage for hash {root_hash:?}, returning empty key set");
-                return Ok(HashMap::new());
-            }
-        };
-
-        // Traverse the tree to collect all keys
-        let mut result = HashMap::new();
-        self.collect_keys_recursive(&root_node, &mut result)?;
-
-        Ok(result)
-    }
-
-    /// Recursively collect keys from a node and its children
-    fn collect_keys_recursive(
-        &self,
-        node: &crate::node::ProllyNode<N>,
-        result: &mut HashMap<Vec<u8>, Vec<u8>>,
-    ) -> Result<(), GitKvError> {
-        if node.is_leaf {
-            // Leaf node: add all key-value pairs
-            for (key, value) in node.keys.iter().zip(node.values.iter()) {
-                result.insert(key.clone(), value.clone());
-            }
-        } else {
-            // Internal node: recursively visit children
-            for value in &node.values {
-                // Value contains the hash of the child node
-                if value.len() == N {
-                    let mut hash_array = [0u8; N];
-                    hash_array.copy_from_slice(value);
-                    let child_hash = ValueDigest(hash_array);
-
-                    if let Some(child_node) = self.tree.storage.get_node_by_hash(&child_hash) {
-                        self.collect_keys_recursive(&child_node, result)?;
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Get commit history for all storage types using Git
-    fn get_commit_history_generic(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        let mut commit_infos = Vec::new();
-
-        // Get HEAD commit
-        let mut head_ref = self
-            .git_repo
-            .head_ref()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get HEAD: {e}")))?
-            .ok_or_else(|| GitKvError::GitObjectError("HEAD not found".to_string()))?;
-
-        // Peel the reference to get the commit ID
-        let peeled_head = head_ref
-            .peel_to_id_in_place()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to peel HEAD: {e}")))?;
-        let mut current_commit_id = peeled_head.detach();
-
-        // Walk through the commit history
-        loop {
-            let mut commit_buffer = Vec::new();
-            let commit_obj = self
-                .git_repo
-                .objects
-                .find(&current_commit_id, &mut commit_buffer)
-                .map_err(|e| GitKvError::GitObjectError(format!("Failed to find commit: {e}")))?;
-
-            let commit = match commit_obj.kind {
-                gix::object::Kind::Commit => gix::objs::CommitRef::from_bytes(commit_obj.data)
-                    .map_err(|e| {
-                        GitKvError::GitObjectError(format!("Failed to parse commit: {e}"))
-                    })?,
-                _ => break,
-            };
-
-            // Create CommitInfo
-            let commit_info = CommitInfo {
-                id: current_commit_id,
-                author: commit.author().name.to_string(),
-                committer: commit.committer().name.to_string(),
-                message: String::from_utf8_lossy(commit.message).to_string(),
-                timestamp: commit.author().time.seconds,
-            };
-
-            commit_infos.push(commit_info);
-
-            // Move to parent commit
-            if let Some(parent_id) = commit.parents.first() {
-                if let Ok(parent_oid) = gix::ObjectId::from_hex(parent_id) {
-                    current_commit_id = parent_oid;
-                } else {
-                    break;
-                }
-            } else {
-                break;
-            }
-        }
-
-        Ok(commit_infos)
-    }
-
-    /// Generic implementation for get_commits_for_key that works with all storage types
-    fn get_commits_for_key_generic(&self, key: &[u8]) -> Result<Vec<CommitInfo>, GitKvError> {
-        let mut commit_history = self.get_commit_history_generic()?;
-
-        // Reverse to process in chronological order (oldest first)
-        commit_history.reverse();
-
-        let mut commits_with_key_changes = Vec::new();
-        let mut previous_value: Option<Vec<u8>> = None; // None = key not present, Some(val) = key present with value
-
-        for commit in commit_history {
-            // Get the key value at this commit by reconstructing tree state
-            let current_value = {
-                if let Ok(tree_config) = self.read_tree_config_from_commit(&commit.id) {
-                    if let Ok(keys_at_commit) = self.collect_keys_from_config(&tree_config) {
-                        keys_at_commit.get(key).cloned()
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            };
-
-            // Check if the value changed from the previous commit
-            let value_changed = previous_value != current_value;
-
-            if value_changed {
-                commits_with_key_changes.push(commit);
-            }
-
-            previous_value = current_value;
-        }
-
-        // Reverse back to newest first for the final result
-        commits_with_key_changes.reverse();
-
-        Ok(commits_with_key_changes)
-    }
-
-    // ============================================================================
-    // Generic checkout, merge, try_merge implementations for all storage backends
-    // All backends use Git for version control, so these operations work the same
-    // ============================================================================
-
-    /// Get commit ID for a branch (generic version)
-    fn get_branch_commit_generic(&self, branch: &str) -> Result<gix::ObjectId, GitKvError> {
-        let branch_ref = format!("refs/heads/{branch}");
-        match self.git_repo.refs.find(&branch_ref) {
-            Ok(reference) => match reference.target.try_id() {
-                Some(commit_id) => Ok(commit_id.to_owned()),
-                None => Err(GitKvError::GitObjectError(format!(
-                    "Branch {branch} does not point to a commit"
-                ))),
-            },
-            Err(_) => Err(GitKvError::BranchNotFound(branch.to_string())),
-        }
-    }
-
-    /// Get parents of a commit (generic version)
-    fn get_commit_parents_generic(
-        &self,
-        commit_id: &gix::ObjectId,
-    ) -> Result<Vec<gix::ObjectId>, GitKvError> {
-        let mut buffer = Vec::new();
-        let commit = self
-            .git_repo
-            .objects
-            .find(commit_id, &mut buffer)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to find commit: {e}")))?;
-
-        let commit_ref = commit
-            .decode()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to decode commit: {e}")))?
-            .into_commit()
-            .ok_or_else(|| GitKvError::GitObjectError("Object is not a commit".to_string()))?;
-
-        Ok(commit_ref.parents().collect())
-    }
-
-    /// Find the merge base (common ancestor) of two branches (generic version)
-    fn find_merge_base_generic(
-        &self,
-        branch1: &str,
-        branch2: &str,
-    ) -> Result<gix::ObjectId, GitKvError> {
-        let commit1 = self.get_branch_commit_generic(branch1)?;
-        let commit2 = self.get_branch_commit_generic(branch2)?;
-
-        let mut visited1 = std::collections::HashSet::new();
-        let mut queue1 = std::collections::VecDeque::new();
-        queue1.push_back(commit1);
-
-        while let Some(commit_id) = queue1.pop_front() {
-            if visited1.contains(&commit_id) {
-                continue;
-            }
-            visited1.insert(commit_id);
-
-            if let Ok(parents) = self.get_commit_parents_generic(&commit_id) {
-                for parent in parents {
-                    if !visited1.contains(&parent) {
-                        queue1.push_back(parent);
-                    }
-                }
-            }
-        }
-
-        let mut visited2 = std::collections::HashSet::new();
-        let mut queue2 = std::collections::VecDeque::new();
-        queue2.push_back(commit2);
-
-        while let Some(commit_id) = queue2.pop_front() {
-            if visited2.contains(&commit_id) {
-                continue;
-            }
-            visited2.insert(commit_id);
-
-            if visited1.contains(&commit_id) {
-                return Ok(commit_id);
-            }
-
-            if let Ok(parents) = self.get_commit_parents_generic(&commit_id) {
-                for parent in parents {
-                    if !visited2.contains(&parent) {
-                        queue2.push_back(parent);
-                    }
-                }
-            }
-        }
-
-        Err(GitKvError::GitObjectError(
-            "No common ancestor found".to_string(),
-        ))
-    }
-
-    /// Reload the tree state from the current HEAD commit (generic version)
-    fn reload_tree_from_head_generic(&mut self) -> Result<(), GitKvError>
-    where
-        Self: HistoricalAccess<N>,
-    {
-        // Get the current HEAD commit
-        let head = self
-            .git_repo
-            .head()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to get HEAD: {e}")))?;
-
-        let head_commit_id = head.id().ok_or_else(|| {
-            GitKvError::GitObjectError("HEAD does not point to a commit".to_string())
-        })?;
-
-        let head_object_id = head_commit_id.detach();
-
-        // Load all key-value pairs from the HEAD commit using HistoricalAccess
-        let keys_at_head = self.collect_keys_from_commit_generic(&head_object_id)?;
-
-        // Get the config from the commit
-        let config = self.read_tree_config_from_commit(&head_object_id)?;
-
-        // Clear and rebuild the tree
-        self.tree.config = config;
-
-        // Clear existing data and insert all keys from HEAD
-        // Note: This creates a new tree with the same storage
-        let storage = self.tree.storage.clone();
-        let tree_config = self.tree.config.clone();
-        self.tree = ProllyTree::new(storage, tree_config);
-
-        for (key, value) in keys_at_head {
-            self.tree.insert(key, value);
-        }
-
-        Ok(())
-    }
-
-    /// Collect all key-value pairs from a specific commit (generic version)
-    fn collect_keys_from_commit_generic(
-        &self,
-        commit_id: &gix::ObjectId,
-    ) -> Result<HashMap<Vec<u8>, Vec<u8>>, GitKvError> {
-        // Read the tree config from the commit
-        let tree_config = self.read_tree_config_from_commit(commit_id)?;
-
-        // Use the generic collect_keys_from_config which works for all storage types
-        self.collect_keys_from_config(&tree_config)
-    }
-
-    /// Switch to a different branch or commit (generic version for all backends)
-    pub fn checkout_generic(&mut self, branch_or_commit: &str) -> Result<(), GitKvError>
-    where
-        Self: HistoricalAccess<N>,
-    {
-        // Clear staging area
-        self.staging_area.clear();
-        self.save_staging_area()?;
-
-        // Update HEAD to point to the new branch/commit
-        let target_ref = if branch_or_commit.starts_with("refs/") {
-            branch_or_commit.to_string()
-        } else {
-            format!("refs/heads/{branch_or_commit}")
-        };
-
-        // Check if the reference exists
-        match self.git_repo.refs.find(&target_ref) {
-            Ok(_reference) => {
-                self.current_branch = branch_or_commit.to_string();
-
-                // Update HEAD to point to the new branch
-                let head_file = self.git_repo.path().join("HEAD");
-                let head_content = format!("ref: refs/heads/{branch_or_commit}");
-                std::fs::write(&head_file, head_content).map_err(|e| {
-                    GitKvError::GitObjectError(format!("Failed to update HEAD: {e}"))
-                })?;
-            }
-            Err(_) => {
-                return Err(GitKvError::BranchNotFound(branch_or_commit.to_string()));
-            }
-        }
-
-        // Reload the tree from the HEAD commit
-        self.reload_tree_from_head_generic()?;
-
-        Ok(())
-    }
-
-    /// Merge another branch into the current branch (generic version for all backends)
-    pub fn merge_generic<R: ConflictResolver>(
-        &mut self,
-        source_branch: &str,
-        resolver: &R,
-    ) -> Result<gix::ObjectId, GitKvError>
-    where
-        Self: HistoricalAccess<N>,
-    {
-        let dest_branch = self.current_branch.clone();
-
-        // Find common base commit
-        let base_commit = self.find_merge_base_generic(&dest_branch, source_branch)?;
-
-        // Get key-value data from each state
-        let base_kv = self.collect_keys_from_commit_generic(&base_commit)?;
-        let source_commit = self.get_branch_commit_generic(source_branch)?;
-        let source_kv = self.collect_keys_from_commit_generic(&source_commit)?;
-        let mut dest_kv = HashMap::new();
-
-        for key in self.tree.collect_keys() {
-            if let Some(value) = self.get(&key) {
-                dest_kv.insert(key, value);
-            }
-        }
-
-        // Perform three-way merge at key-value level
-        let mut merge_results = Vec::new();
-        let mut all_keys = std::collections::HashSet::new();
-
-        all_keys.extend(base_kv.keys().cloned());
-        all_keys.extend(source_kv.keys().cloned());
-        all_keys.extend(dest_kv.keys().cloned());
-
-        for key in all_keys {
-            let base_value = base_kv.get(&key);
-            let source_value = source_kv.get(&key);
-            let dest_value = dest_kv.get(&key);
-
-            match (base_value, source_value, dest_value) {
-                // Key exists in all three - check for modifications
-                (Some(base), Some(source), Some(dest)) => {
-                    if base == source && base == dest {
-                        // No changes, skip
-                        continue;
-                    } else if base == dest && base != source {
-                        // Only source changed - take source value
-                        merge_results.push(crate::diff::MergeResult::Modified(
-                            key.clone(),
-                            source.clone(),
-                        ));
-                    } else if base == source && base != dest {
-                        // Only dest changed - keep dest value (no-op)
-                        continue;
-                    } else if source == dest {
-                        // Both changed to same value - keep it (no-op)
-                        continue;
-                    } else {
-                        // Conflict: both branches modified differently
-                        let conflict = crate::diff::MergeConflict {
-                            key: key.clone(),
-                            base_value: Some(base.clone()),
-                            source_value: Some(source.clone()),
-                            destination_value: Some(dest.clone()),
-                        };
-                        merge_results.push(crate::diff::MergeResult::Conflict(conflict));
-                    }
-                }
-                // Key added in source, not in base or dest
-                (None, Some(source), None) => {
-                    merge_results
-                        .push(crate::diff::MergeResult::Added(key.clone(), source.clone()));
-                }
-                // Key added in dest, not in base or source - keep it (no-op)
-                (None, None, Some(_dest)) => {
-                    continue;
-                }
-                // Key added in both source and dest - potential conflict
-                (None, Some(source), Some(dest)) => {
-                    if source == dest {
-                        // Both added same value - keep it (no-op)
-                        continue;
-                    } else {
-                        // Conflict: both branches added different values
-                        let conflict = crate::diff::MergeConflict {
-                            key: key.clone(),
-                            base_value: None,
-                            source_value: Some(source.clone()),
-                            destination_value: Some(dest.clone()),
-                        };
-                        merge_results.push(crate::diff::MergeResult::Conflict(conflict));
-                    }
-                }
-                // Key deleted in source, still exists in dest
-                (Some(_base), None, Some(_dest)) => {
-                    // Source deleted it - apply deletion
-                    merge_results.push(crate::diff::MergeResult::Removed(key.clone()));
-                }
-                // Key deleted in dest, still exists in source - keep deletion (no-op)
-                (Some(_base), Some(_source), None) => {
-                    continue;
-                }
-                // Key deleted in both - no-op
-                (Some(_base), None, None) => {
-                    continue;
-                }
-                // All other cases - no action needed
-                _ => continue,
-            }
-        }
-
-        // Apply conflict resolution
-        let mut resolved_results = Vec::new();
-        let mut unresolved_conflicts = Vec::new();
-
-        for result in merge_results {
-            match result {
-                crate::diff::MergeResult::Conflict(conflict) => {
-                    if let Some(resolved_result) = resolver.resolve_conflict(&conflict) {
-                        resolved_results.push(resolved_result);
-                    } else {
-                        unresolved_conflicts.push(conflict);
-                    }
-                }
-                other => resolved_results.push(other),
-            }
-        }
-
-        // If there are unresolved conflicts, return them
-        if !unresolved_conflicts.is_empty() {
-            return Err(GitKvError::MergeConflictError(unresolved_conflicts));
-        }
-
-        // Apply resolved merge results directly to current tree
-        for result in resolved_results {
-            match result {
-                crate::diff::MergeResult::Added(key, value) => {
-                    self.tree.insert(key, value);
-                }
-                crate::diff::MergeResult::Removed(key) => {
-                    self.tree.delete(&key);
-                }
-                crate::diff::MergeResult::Modified(key, value) => {
-                    self.tree.insert(key, value);
-                }
-                crate::diff::MergeResult::Conflict(_) => {
-                    // Should not happen after resolution
-                }
-            }
-        }
-
-        // Persist tree changes
-        self.tree.persist_root();
-        self.tree
-            .save_config()
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to save config: {e}")))?;
-
-        self.save_tree_config_to_git_internal()?;
-
-        // Commit the merge
-        let message = format!("Merge branch '{}' into {}", source_branch, dest_branch);
-        let commit_id = self.commit(&message)?;
-
-        Ok(commit_id)
-    }
-
-    /// Try to merge another branch and return conflicts (generic version)
-    ///
-    /// This method attempts to merge the source branch into the current branch.
-    /// If there are no conflicts, the merge is applied and committed.
-    /// If there are conflicts, the merge is not applied and the conflicts are returned.
-    pub fn try_merge_generic(&mut self, source_branch: &str) -> Result<gix::ObjectId, GitKvError>
-    where
-        Self: HistoricalAccess<N> + TreeConfigSaver<N>,
-    {
-        // Use a NoOpResolver that leaves all conflicts unresolved
-        struct NoOpResolver;
-        impl crate::diff::ConflictResolver for NoOpResolver {
-            fn resolve_conflict(
-                &self,
-                _conflict: &crate::diff::MergeConflict,
-            ) -> Option<crate::diff::MergeResult> {
-                None // Leave all conflicts unresolved
-            }
-        }
-
-        self.merge_generic(source_branch, &NoOpResolver)
-    }
-}
-
-// Implement TreeConfigSaver for InMemoryNodeStorage
-impl<const N: usize> TreeConfigSaver<N> for VersionedKvStore<N, InMemoryNodeStorage<N>> {
-    fn save_tree_config_to_git_internal(&self) -> Result<(), GitKvError> {
-        self.save_tree_config_to_git()
-    }
-}
-
-// Specialized implementation for InMemoryNodeStorage
-impl<const N: usize> VersionedKvStore<N, InMemoryNodeStorage<N>> {
-    /// Save tree config to git for InMemoryNodeStorage
-    ///
-    /// Writes only the config (with root hash) to dataset directory.
-    /// The config is committed to git for historical access.
-    /// Nodes are kept in memory - historical access works within the same session
-    /// by loading nodes from the in-memory storage using the root hash.
-    /// After restart, nodes are lost (expected for in-memory storage).
-    fn save_tree_config_to_git(&self) -> Result<(), GitKvError> {
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".to_string()))?;
-
-        // Get the current tree configuration (includes root_hash)
-        let config = self.tree.config.clone();
-
-        // Serialize the config to JSON
-        let config_json = serde_json::to_string_pretty(&config)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to serialize config: {e}")))?;
-
-        // Write only the config file to the dataset directory
-        // No need for prolly_hash_mappings - nodes are in memory and accessed by root_hash
-        // Historical access uses root_hash to traverse the tree from in-memory storage
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        std::fs::write(&config_path, config_json)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to write config file: {e}")))?;
-
-        Ok(())
-    }
-}
-
-// Implement TreeConfigSaver for FileNodeStorage
-impl<const N: usize> TreeConfigSaver<N> for VersionedKvStore<N, FileNodeStorage<N>> {
-    fn save_tree_config_to_git_internal(&self) -> Result<(), GitKvError> {
-        self.save_tree_config_to_git()
-    }
-}
-
-// Specialized implementation for FileNodeStorage
-impl<const N: usize> VersionedKvStore<N, FileNodeStorage<N>> {
-    /// Save tree config to git for FileNodeStorage
-    ///
-    /// Writes only the config (with root hash) to dataset directory.
-    /// The config is committed to git for historical access.
-    /// Nodes remain in .git/prolly/nodes/files/ (shared, content-addressed).
-    /// Historical access reconstructs tree by loading nodes from storage using root hash.
-    fn save_tree_config_to_git(&self) -> Result<(), GitKvError> {
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".to_string()))?;
-
-        // Get the current tree configuration (includes root_hash)
-        let config = self.tree.config.clone();
-
-        // Serialize the config to JSON
-        let config_json = serde_json::to_string_pretty(&config)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to serialize config: {e}")))?;
-
-        // Write only the config file to the dataset directory
-        // No need for prolly_hash_mappings - nodes are content-addressed in .git/prolly/nodes/files/
-        // Historical access uses root_hash to traverse the tree from node storage
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        std::fs::write(&config_path, config_json)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to write config file: {e}")))?;
-
-        Ok(())
-    }
-}
-
-// Implement TreeConfigSaver for RocksDBNodeStorage
-#[cfg(feature = "rocksdb_storage")]
-impl<const N: usize> TreeConfigSaver<N> for VersionedKvStore<N, RocksDBNodeStorage<N>> {
-    fn save_tree_config_to_git_internal(&self) -> Result<(), GitKvError> {
-        self.save_tree_config_to_git()
-    }
-}
-
-// Specialized implementation for RocksDBNodeStorage
-#[cfg(feature = "rocksdb_storage")]
-impl<const N: usize> VersionedKvStore<N, RocksDBNodeStorage<N>> {
-    /// Save tree config to git for RocksDBNodeStorage
-    ///
-    /// Writes only the config (with root hash) to dataset directory.
-    /// The config is committed to git for historical access.
-    /// Nodes remain in .git/prolly/nodes/rocksdb/ (shared, content-addressed).
-    /// Historical access reconstructs tree by loading nodes from storage using root hash.
-    fn save_tree_config_to_git(&self) -> Result<(), GitKvError> {
-        let dataset_dir = self
-            .dataset_dir
-            .as_ref()
-            .ok_or_else(|| GitKvError::GitObjectError("Dataset directory not set".to_string()))?;
-
-        // Get the current tree configuration (includes root_hash)
-        let config = self.tree.config.clone();
-
-        // Serialize the config to JSON
-        let config_json = serde_json::to_string_pretty(&config)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to serialize config: {e}")))?;
-
-        // Write only the config file to the dataset directory
-        // No need for prolly_hash_mappings - nodes are content-addressed in .git/prolly/nodes/rocksdb/
-        // Historical access uses root_hash to traverse the tree from node storage
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        std::fs::write(&config_path, config_json)
-            .map_err(|e| GitKvError::GitObjectError(format!("Failed to write config file: {e}")))?;
-
-        Ok(())
-    }
-}
-
-// ==============================================================================
-// Thread-Safe Wrapper Implementation
-// ==============================================================================
-
-impl<const N: usize> ThreadSafeGitVersionedKvStore<N> {
-    /// Initialize a new thread-safe git-backed versioned key-value store
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = GitVersionedKvStore::init(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-
-    /// Open an existing thread-safe git-backed versioned key-value store
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = GitVersionedKvStore::open(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-}
-
-impl<const N: usize, S: NodeStorage<N>> ThreadSafeVersionedKvStore<N, S>
-where
-    VersionedKvStore<N, S>: TreeConfigSaver<N>,
-{
-    /// Insert a key-value pair (stages the change)
-    pub fn insert(&self, key: Vec<u8>, value: Vec<u8>) -> Result<(), GitKvError> {
-        let mut store = self.inner.lock();
-        store.insert(key, value)
-    }
-
-    /// Update an existing key-value pair (stages the change)
-    pub fn update(&self, key: Vec<u8>, value: Vec<u8>) -> Result<bool, GitKvError> {
-        let mut store = self.inner.lock();
-        store.update(key, value)
-    }
-
-    /// Delete a key-value pair (stages the change)
-    pub fn delete(&self, key: &[u8]) -> Result<bool, GitKvError> {
-        let mut store = self.inner.lock();
-        store.delete(key)
-    }
-
-    /// Get a value by key (checks staging area first, then committed data)
-    pub fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
-        let store = self.inner.lock();
-        store.get(key)
-    }
-
-    /// List all keys (includes staged changes)
-    pub fn list_keys(&self) -> Result<Vec<Vec<u8>>, GitKvError> {
-        let store = self.inner.lock();
-        Ok(store.list_keys())
-    }
-
-    /// Show current staging area status
-    pub fn status(&self) -> Result<Vec<(Vec<u8>, String)>, GitKvError> {
-        let store = self.inner.lock();
-        Ok(store.status())
-    }
-
-    /// Commit staged changes
-    pub fn commit(&self, message: &str) -> Result<gix::ObjectId, GitKvError> {
-        let mut store = self.inner.lock();
-        store.commit(message)
-    }
-
-    /// Create a new branch
-    pub fn create_branch(&self, name: &str) -> Result<(), GitKvError> {
-        let mut store = self.inner.lock();
-        store.create_branch(name)
-    }
-
-    /// Get commit history
-    pub fn log(&self) -> Result<Vec<CommitInfo>, GitKvError> {
-        let store = self.inner.lock();
-        store.log()
-    }
-
-    /// Get current branch name
-    pub fn current_branch(&self) -> Result<String, GitKvError> {
-        let store = self.inner.lock();
-        Ok(store.current_branch().to_string())
-    }
-
-    /// Get the underlying git repository reference (creates a clone)
-    pub fn git_repo(&self) -> Result<gix::Repository, GitKvError> {
-        let store = self.inner.lock();
-        Ok(store.git_repo().clone())
-    }
-}
-
-impl<const N: usize> ThreadSafeGitVersionedKvStore<N> {
-    /// Switch to a different branch - Git-specific implementation
-    pub fn checkout(&self, name: &str) -> Result<(), GitKvError> {
-        let mut store = self.inner.lock();
-        store.checkout(name)
-    }
-}
-
-impl<const N: usize> ThreadSafeInMemoryVersionedKvStore<N> {
-    /// Initialize a new thread-safe in-memory versioned key-value store
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = VersionedKvStore::<N, InMemoryNodeStorage<N>>::init(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-
-    /// Open an existing thread-safe in-memory versioned key-value store
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = VersionedKvStore::<N, InMemoryNodeStorage<N>>::open(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-}
-
-impl<const N: usize> ThreadSafeFileVersionedKvStore<N> {
-    /// Initialize a new thread-safe file-based versioned key-value store
-    pub fn init<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = VersionedKvStore::<N, FileNodeStorage<N>>::init(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-
-    /// Open an existing thread-safe file-based versioned key-value store
-    pub fn open<P: AsRef<Path>>(path: P) -> Result<Self, GitKvError> {
-        let inner = VersionedKvStore::<N, FileNodeStorage<N>>::open(path)?;
-        Ok(Self {
-            inner: Arc::new(Mutex::new(inner)),
-        })
-    }
-}
-
-impl<const N: usize, S: NodeStorage<N>> Clone for ThreadSafeVersionedKvStore<N, S> {
-    fn clone(&self) -> Self {
-        Self {
-            inner: Arc::clone(&self.inner),
-        }
-    }
-}
-
-// Implement Send and Sync for the thread-safe wrapper
-unsafe impl<const N: usize, S: NodeStorage<N>> Send for ThreadSafeVersionedKvStore<N, S> where
-    S: Send
-{
-}
-unsafe impl<const N: usize, S: NodeStorage<N>> Sync for ThreadSafeVersionedKvStore<N, S> where
-    S: Send
-{
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_versioned_store_init() {
-        let temp_dir = TempDir::new().unwrap();
-        // Initialize git repository (regular, not bare)
-        gix::init(temp_dir.path()).unwrap();
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-        let store = GitVersionedKvStore::<32>::init(&dataset_dir);
-        assert!(store.is_ok());
-    }
-
-    #[test]
-    fn test_basic_kv_operations() {
-        let temp_dir = TempDir::new().unwrap();
-        // Initialize git repository (regular, not bare)
-        gix::init(temp_dir.path()).unwrap();
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Test insert and get
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        assert_eq!(store.get(b"key1"), Some(b"value1".to_vec()));
-
-        // Test update
-        store
-            .update(b"key1".to_vec(), b"new_value1".to_vec())
-            .unwrap();
-        assert_eq!(store.get(b"key1"), Some(b"new_value1".to_vec()));
-
-        // Test delete
-        store.delete(b"key1").unwrap();
-        assert_eq!(store.get(b"key1"), None);
-    }
-
-    #[test]
-    fn test_commit_workflow() {
-        let temp_dir = TempDir::new().unwrap();
-        // Initialize git repository (regular, not bare)
-        gix::init(temp_dir.path()).unwrap();
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Stage changes
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-
-        // Check status
-        let status = store.status();
-        assert_eq!(status.len(), 2);
-
-        // Commit
-        let commit_id = store.commit("Add initial data").unwrap();
-        // Now we have a real implementation that returns valid commit IDs
-        assert!(!commit_id.is_null());
-
-        // Check that staging area is clear
-        let status = store.status();
-        assert_eq!(status.len(), 0);
-    }
-
-    #[test]
-    fn test_single_commit_behavior() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Get initial commit count
-        let log_output = std::process::Command::new("git")
-            .args(&["log", "--oneline"])
-            .current_dir(temp_dir.path())
-            .output()
-            .unwrap();
-        let initial_commits = String::from_utf8_lossy(&log_output.stdout).lines().count();
-
-        // Insert some data and commit
-        store
-            .insert(b"test_key".to_vec(), b"test_value".to_vec())
-            .unwrap();
-        store.commit("Test single commit").unwrap();
-
-        // Get commit count after our commit
-        let log_output = std::process::Command::new("git")
-            .args(&["log", "--oneline"])
-            .current_dir(temp_dir.path())
-            .output()
-            .unwrap();
-        let final_commits = String::from_utf8_lossy(&log_output.stdout).lines().count();
-
-        // Should have exactly one more commit (no separate metadata commit)
-        assert_eq!(
-            final_commits,
-            initial_commits + 1,
-            "Expected exactly one new commit, but got {} new commits",
-            final_commits - initial_commits
-        );
-
-        // Verify the prolly metadata files exist in the dataset directory
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        let mapping_path = dataset_dir.join("prolly_hash_mappings");
-        assert!(
-            config_path.exists(),
-            "prolly_config_tree_config should exist"
-        );
-        assert!(mapping_path.exists(), "prolly_hash_mappings should exist");
-    }
-
-    #[test]
-    fn test_diff_between_commits() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create first commit with some data
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-        let commit1 = store.commit("Initial data").unwrap();
-
-        // Create second commit with modifications
-        store
-            .update(b"key1".to_vec(), b"value1_modified".to_vec())
-            .unwrap();
-        store.insert(b"key3".to_vec(), b"value3".to_vec()).unwrap();
-        store.delete(b"key2").unwrap();
-        let commit2 = store.commit("Modify data").unwrap();
-
-        // Diff between the two commits
-        let diffs = store
-            .diff(&commit1.to_hex().to_string(), &commit2.to_hex().to_string())
-            .unwrap();
-
-        println!("Diffs found: {}", diffs.len());
-        for diff in &diffs {
-            println!(
-                "  {:?}: {:?}",
-                String::from_utf8_lossy(&diff.key),
-                diff.operation
-            );
-        }
-
-        // Should have 3 changes: key1 modified, key2 removed, key3 added
-        assert_eq!(diffs.len(), 3);
-
-        // Check each diff (they are sorted by key)
-        assert_eq!(diffs[0].key, b"key1");
-        match &diffs[0].operation {
-            DiffOperation::Modified { old, new } => {
-                assert_eq!(old, b"value1");
-                assert_eq!(new, b"value1_modified");
-            }
-            _ => panic!("Expected key1 to be modified"),
-        }
-
-        assert_eq!(diffs[1].key, b"key2");
-        match &diffs[1].operation {
-            DiffOperation::Removed(value) => {
-                assert_eq!(value, b"value2");
-            }
-            _ => panic!("Expected key2 to be removed"),
-        }
-
-        assert_eq!(diffs[2].key, b"key3");
-        match &diffs[2].operation {
-            DiffOperation::Added(value) => {
-                assert_eq!(value, b"value3");
-            }
-            _ => panic!("Expected key3 to be added"),
-        }
-    }
-
-    #[test]
-    fn test_diff_between_branches() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create initial commit on main branch
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-        store.commit("Initial data").unwrap();
-
-        // Create and switch to feature branch
-        store.create_branch("feature").unwrap();
-
-        // Make changes on feature branch
-        store
-            .update(b"key1".to_vec(), b"value1_feature".to_vec())
-            .unwrap();
-        store.insert(b"key3".to_vec(), b"value3".to_vec()).unwrap();
-        store.commit("Feature changes").unwrap();
-
-        // Diff between main and feature branches
-        let diffs = store.diff("main", "feature").unwrap();
-
-        // Should have 2 changes: key1 modified, key3 added
-        assert_eq!(diffs.len(), 2);
-
-        assert_eq!(diffs[0].key, b"key1");
-        match &diffs[0].operation {
-            DiffOperation::Modified { old, new } => {
-                assert_eq!(old, b"value1");
-                assert_eq!(new, b"value1_feature");
-            }
-            _ => panic!("Expected key1 to be modified"),
-        }
-
-        assert_eq!(diffs[1].key, b"key3");
-        match &diffs[1].operation {
-            DiffOperation::Added(value) => {
-                assert_eq!(value, b"value3");
-            }
-            _ => panic!("Expected key3 to be added"),
-        }
-    }
-
-    #[test]
-    fn test_init_with_existing_store() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        // First init - creates new store
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Add some data and commit
-        store
-            .insert(b"test_key".to_vec(), b"test_value".to_vec())
-            .unwrap();
-        store.commit("Test data").unwrap();
-
-        // Drop the store to ensure we're testing a fresh init
-        drop(store);
-
-        // Second init - should load existing store, not create new one
-        let store2 = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Verify data still exists (wasn't overwritten)
-        assert_eq!(
-            store2.get(b"test_key"),
-            Some(b"test_value".to_vec()),
-            "Data should still exist after second init"
-        );
-
-        // Verify config files exist
-        let config_path = dataset_dir.join("prolly_config_tree_config");
-        let mapping_path = dataset_dir.join("prolly_hash_mappings");
-        assert!(config_path.exists(), "Config file should exist");
-        assert!(mapping_path.exists(), "Mapping file should exist");
-    }
-
-    #[test]
-    fn test_diff_with_no_changes() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create a commit
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        let commit = store.commit("Initial data").unwrap();
-
-        // Diff the commit with itself
-        let diffs = store
-            .diff(&commit.to_hex().to_string(), &commit.to_hex().to_string())
-            .unwrap();
-
-        // Should have no changes
-        assert_eq!(diffs.len(), 0);
-    }
-
-    #[test]
-    fn test_diff_with_inmemory_storage() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = InMemoryVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Add some data and create first commit
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-        let commit1 = store.commit("Initial data").unwrap();
-
-        // Make changes and create second commit
-        store
-            .update(b"key1".to_vec(), b"updated_value1".to_vec())
-            .unwrap();
-        store.insert(b"key3".to_vec(), b"value3".to_vec()).unwrap();
-        let commit2 = store.commit("Update data").unwrap();
-
-        // Test diff between the two commits - should now work with actual git references
-        let diffs = store
-            .diff(&commit1.to_hex().to_string(), &commit2.to_hex().to_string())
-            .unwrap();
-
-        // Should have 2 changes: key1 modified, key3 added
-        assert_eq!(diffs.len(), 2);
-
-        // Test diff with HEAD (should compare commit1 to current HEAD)
-        let head_diffs = store.diff(&commit1.to_hex().to_string(), "HEAD").unwrap();
-        assert_eq!(head_diffs.len(), 2);
-
-        // Test diff with same commit (should have no changes)
-        let same_diffs = store
-            .diff(&commit1.to_hex().to_string(), &commit1.to_hex().to_string())
-            .unwrap();
-        assert_eq!(same_diffs.len(), 0);
-    }
-
-    #[test]
-    fn test_get_commits_for_key() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create commit 1: Add key1
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-        let commit1 = store.commit("Add key1 and key2").unwrap();
-
-        // Create commit 2: Modify key1, leave key2 unchanged
-        store
-            .update(b"key1".to_vec(), b"value1_modified".to_vec())
-            .unwrap();
-        let commit2 = store.commit("Modify key1").unwrap();
-
-        // Create commit 3: Add key3, leave key1 and key2 unchanged
-        store.insert(b"key3".to_vec(), b"value3".to_vec()).unwrap();
-        let commit3 = store.commit("Add key3").unwrap();
-
-        // Create commit 4: Delete key1
-        store.delete(b"key1").unwrap();
-        let commit4 = store.commit("Delete key1").unwrap();
-
-        // Test get_commits for key1 (should have commits 4, 2, 1 - newest first)
-        let key1_commits = store.get_commits(b"key1").unwrap();
-
-        // Debug: print commit information
-        eprintln!("key1_commits found: {}", key1_commits.len());
-        for (i, commit) in key1_commits.iter().enumerate() {
-            eprintln!("  [{}] {} - {}", i, commit.id, commit.message.trim());
-        }
-        eprintln!("Expected commits:");
-        eprintln!("  commit4 (delete): {}", commit4);
-        eprintln!("  commit2 (modify): {}", commit2);
-        eprintln!("  commit1 (add): {}", commit1);
-
-        assert_eq!(key1_commits.len(), 3);
-        assert_eq!(key1_commits[0].id, commit4); // Delete commit
-        assert_eq!(key1_commits[1].id, commit2); // Modify commit
-        assert_eq!(key1_commits[2].id, commit1); // Add commit
-
-        // Test get_commits for key2 (should have only commit 1)
-        let key2_commits = store.get_commits(b"key2").unwrap();
-        assert_eq!(key2_commits.len(), 1);
-        assert_eq!(key2_commits[0].id, commit1); // Add commit
-
-        // Test get_commits for key3 (should have only commit 3)
-        let key3_commits = store.get_commits(b"key3").unwrap();
-        assert_eq!(key3_commits.len(), 1);
-        assert_eq!(key3_commits[0].id, commit3); // Add commit
-
-        // Test get_commits for non-existent key (should be empty)
-        let nonexistent_commits = store.get_commits(b"nonexistent").unwrap();
-        assert_eq!(nonexistent_commits.len(), 0);
-    }
-
-    #[test]
-    fn test_get_commits_with_repeated_changes() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create commit 1: Add key
-        store.insert(b"key".to_vec(), b"value1".to_vec()).unwrap();
-        let commit1 = store.commit("Add key with value1").unwrap();
-
-        // Create commit 2: Change key to same value (should not be tracked)
-        store.update(b"key".to_vec(), b"value1".to_vec()).unwrap();
-        let _commit2 = store.commit("Update key to same value").unwrap();
-
-        // Create commit 3: Change key to different value
-        store.update(b"key".to_vec(), b"value2".to_vec()).unwrap();
-        let commit3 = store.commit("Change key to value2").unwrap();
-
-        // Create commit 4: Change key back to original value
-        store.update(b"key".to_vec(), b"value1".to_vec()).unwrap();
-        let commit4 = store.commit("Change key back to value1").unwrap();
-
-        // Test get_commits for key - should have commits 4, 3, 1 (skipping commit2 since no real change)
-        let key_commits = store.get_commits(b"key").unwrap();
-        assert_eq!(key_commits.len(), 3);
-        assert_eq!(key_commits[0].id, commit4); // Back to value1
-        assert_eq!(key_commits[1].id, commit3); // Changed to value2
-        assert_eq!(key_commits[2].id, commit1); // Initial add
-    }
-
-    #[test]
-    fn test_historical_access_non_git_storages() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository
-        gix::init(temp_dir.path()).unwrap();
-
-        // Create subdirectory for dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        // Test InMemory storage
-        {
-            let mut store = InMemoryVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-            // Add some data and commit
-            store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-            store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-            let commit_id = store.commit("Initial data").unwrap();
-
-            // Test historical access
-            // InMemory storage now saves tree config to git commits, enabling full historical functionality
-            let keys_at_head = store.get_keys_at_ref("HEAD").unwrap();
-            assert_eq!(keys_at_head.len(), 2);
-            assert!(keys_at_head.contains_key(&b"key1".to_vec()));
-            assert!(keys_at_head.contains_key(&b"key2".to_vec()));
-
-            // Test access by commit ID
-            let keys_at_commit = store
-                .get_keys_at_ref(&commit_id.to_hex().to_string())
-                .unwrap();
-            assert_eq!(keys_at_commit.len(), 2);
-
-            // Test commit history access - this should work as it only reads git commit metadata
-            let commit_history = store.get_commit_history().unwrap();
-            assert!(!commit_history.is_empty());
-
-            // Test get_commits_for_key - now works with tree config available
-            let key1_commits = store.get_commits(b"key1").unwrap();
-            assert!(!key1_commits.is_empty());
-        }
-
-        // Test File storage
-        {
-            let mut store = FileVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-            // Add some data and commit
-            store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-            store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-            let _commit_id = store.commit("Initial data").unwrap();
-
-            // Test historical access
-            // File storage now saves tree config to git commits, enabling full historical functionality
-            let keys_at_head = store.get_keys_at_ref("HEAD").unwrap();
-            assert_eq!(keys_at_head.len(), 2);
-            assert!(keys_at_head.contains_key(&b"key1".to_vec()));
-            assert!(keys_at_head.contains_key(&b"key2".to_vec()));
-
-            // Test commit history access - this should work
-            let commit_history = store.get_commit_history().unwrap();
-            assert!(!commit_history.is_empty());
-
-            // Test get_commits_for_key - now works with tree config available
-            let key1_commits = store.get_commits(b"key1").unwrap();
-            assert!(!key1_commits.is_empty());
-        }
-
-        // Test RocksDB storage (if enabled)
-        #[cfg(feature = "rocksdb_storage")]
-        {
-            let mut store = RocksDBVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-            // Add some data and commit
-            store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-            store.insert(b"key2".to_vec(), b"value2".to_vec()).unwrap();
-            let _commit_id = store.commit("Initial data").unwrap();
-
-            // Test historical access
-            // RocksDB storage now saves tree config to git commits, enabling full historical functionality
-            let keys_at_head = store.get_keys_at_ref("HEAD").unwrap();
-            assert_eq!(keys_at_head.len(), 2);
-            assert!(keys_at_head.contains_key(&b"key1".to_vec()));
-            assert!(keys_at_head.contains_key(&b"key2".to_vec()));
-
-            // Test commit history access - this should work
-            let commit_history = store.get_commit_history().unwrap();
-            assert!(!commit_history.is_empty());
-
-            // Test get_commits_for_key - now works with tree config available
-            let key1_commits = store.get_commits(b"key1").unwrap();
-            assert!(!key1_commits.is_empty());
-        }
-    }
-
-    #[test]
-    fn test_get_commits_complex_multi_branch_scenarios() {
-        let temp_dir = TempDir::new().unwrap();
-        gix::init(temp_dir.path()).unwrap();
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // === Main branch development ===
-        // Initial commit with key1
-        store
-            .insert(b"key1".to_vec(), b"value1_v1".to_vec())
-            .unwrap();
-        store
-            .insert(b"shared_key".to_vec(), b"shared_v1".to_vec())
-            .unwrap();
-        let commit1 = store
-            .commit("Initial commit with key1 and shared_key")
-            .unwrap();
-
-        // Second commit modifying key1 and adding key2
-        store
-            .update(b"key1".to_vec(), b"value1_v2".to_vec())
-            .unwrap();
-        store
-            .insert(b"key2".to_vec(), b"value2_v1".to_vec())
-            .unwrap();
-        let commit2 = store.commit("Update key1, add key2").unwrap();
-
-        // === Create feature branch ===
-        store.create_branch("feature/new-keys").unwrap();
-        store.checkout("feature/new-keys").unwrap();
-
-        // Branch commit 1: modify key2 and add key3
-        store
-            .update(b"key2".to_vec(), b"value2_branch_v1".to_vec())
-            .unwrap();
-        store
-            .insert(b"key3".to_vec(), b"value3_branch_v1".to_vec())
-            .unwrap();
-        store
-            .update(b"shared_key".to_vec(), b"shared_branch_v1".to_vec())
-            .unwrap();
-        let branch_commit1 = store
-            .commit("Feature branch: modify key2, add key3, update shared_key")
-            .unwrap();
-
-        // Branch commit 2: further modify key3
-        store
-            .update(b"key3".to_vec(), b"value3_branch_v2".to_vec())
-            .unwrap();
-        let branch_commit2 = store.commit("Feature branch: update key3 again").unwrap();
-
-        // === Back to main branch ===
-        store.checkout("main").unwrap();
-
-        // Main commit 3: delete key2, modify shared_key differently
-        store.delete(b"key2").unwrap();
-        store
-            .update(b"shared_key".to_vec(), b"shared_main_v2".to_vec())
-            .unwrap();
-        let main_commit3 = store
-            .commit("Main: delete key2, update shared_key")
-            .unwrap();
-
-        // === Create another branch for testing ===
-        store.create_branch("hotfix/key1-fix").unwrap();
-        store.checkout("hotfix/key1-fix").unwrap();
-
-        // Hotfix: critical update to key1
-        store
-            .update(b"key1".to_vec(), b"value1_hotfixed".to_vec())
-            .unwrap();
-        let hotfix_commit = store.commit("Hotfix: critical key1 update").unwrap();
-
-        // === Test 1: Get commits for key1 across all branches ===
-        println!("\n=== Testing key1 commits across branches ===");
-
-        // Test from main branch perspective
-        store.checkout("main").unwrap();
-        let key1_commits_main = store.get_commits(b"key1").unwrap();
-        println!("Key1 commits from main branch: {}", key1_commits_main.len());
-        for (i, commit) in key1_commits_main.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see: commit2 (update), commit1 (initial) - but not hotfix since we're on main
-        assert_eq!(key1_commits_main.len(), 2);
-        assert_eq!(key1_commits_main[0].id, commit2); // Most recent first
-        assert_eq!(key1_commits_main[1].id, commit1);
-
-        // Test from hotfix branch perspective
-        store.checkout("hotfix/key1-fix").unwrap();
-        let key1_commits_hotfix = store.get_commits(b"key1").unwrap();
-        println!(
-            "Key1 commits from hotfix branch: {}",
-            key1_commits_hotfix.len()
-        );
-        for (i, commit) in key1_commits_hotfix.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see hotfix commit, then main branch history
-        assert_eq!(key1_commits_hotfix.len(), 3);
-        assert_eq!(key1_commits_hotfix[0].id, hotfix_commit);
-        assert_eq!(key1_commits_hotfix[1].id, commit2);
-        assert_eq!(key1_commits_hotfix[2].id, commit1);
-
-        // === Test 2: Get commits for key2 (created then deleted on main, modified on feature) ===
-        println!("\n=== Testing key2 commits across branches ===");
-
-        // From main branch (key2 was deleted)
-        store.checkout("main").unwrap();
-        let key2_commits_main = store.get_commits(b"key2").unwrap();
-        println!("Key2 commits from main branch: {}", key2_commits_main.len());
-        for (i, commit) in key2_commits_main.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see: main_commit3 (delete), commit2 (add)
-        assert_eq!(key2_commits_main.len(), 2);
-        assert_eq!(key2_commits_main[0].id, main_commit3);
-        assert_eq!(key2_commits_main[1].id, commit2);
-
-        // From feature branch (key2 was modified)
-        store.checkout("feature/new-keys").unwrap();
-        let key2_commits_feature = store.get_commits(b"key2").unwrap();
-        println!(
-            "Key2 commits from feature branch: {}",
-            key2_commits_feature.len()
-        );
-        for (i, commit) in key2_commits_feature.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see: branch_commit1 (modify), commit2 (add from main)
-        assert_eq!(key2_commits_feature.len(), 2);
-        assert_eq!(key2_commits_feature[0].id, branch_commit1);
-        assert_eq!(key2_commits_feature[1].id, commit2);
-
-        // === Test 3: Get commits for key3 (only exists on feature branch) ===
-        println!("\n=== Testing key3 commits (feature branch only) ===");
-
-        // From feature branch
-        let key3_commits_feature = store.get_commits(b"key3").unwrap();
-        println!(
-            "Key3 commits from feature branch: {}",
-            key3_commits_feature.len()
-        );
-        for (i, commit) in key3_commits_feature.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see both feature branch commits
-        assert_eq!(key3_commits_feature.len(), 2);
-        assert_eq!(key3_commits_feature[0].id, branch_commit2);
-        assert_eq!(key3_commits_feature[1].id, branch_commit1);
-
-        // From main branch (key3 doesn't exist)
-        store.checkout("main").unwrap();
-
-        // Debug: Let's check what keys exist at HEAD on main
-        let keys_at_main_head = store.get_keys_at_ref("HEAD").unwrap();
-        println!(
-            "Keys at main HEAD: {:?}",
-            keys_at_main_head.keys().collect::<Vec<_>>()
-        );
-        println!(
-            "Key3 value at main HEAD: {:?}",
-            keys_at_main_head.get(&b"key3".to_vec())
-        );
-
-        let key3_commits_main = store.get_commits(b"key3").unwrap();
-        println!("Key3 commits from main branch: {}", key3_commits_main.len());
-        for (i, commit) in key3_commits_main.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-            // Check what keys existed at this specific commit
-            let keys_at_commit = store.collect_keys_at_commit(&commit.id).unwrap();
-            println!(
-                "    Keys at this commit: {:?}",
-                keys_at_commit.keys().collect::<Vec<_>>()
-            );
-            println!(
-                "    Key3 value at this commit: {:?}",
-                keys_at_commit.get(&b"key3".to_vec())
-            );
-        }
-
-        // For now, let's just verify that key3 doesn't exist at the current main HEAD
-        // The issue might be in the commit history logic, but the current state should be correct
-        assert!(
-            !keys_at_main_head.contains_key(&b"key3".to_vec()),
-            "key3 should not exist at main HEAD"
-        );
-
-        // === Test 4: Get commits for shared_key (modified differently on different branches) ===
-        println!("\n=== Testing shared_key commits across branches ===");
-
-        // From main branch
-        let shared_commits_main = store.get_commits(b"shared_key").unwrap();
-        println!(
-            "Shared_key commits from main branch: {}",
-            shared_commits_main.len()
-        );
-        for (i, commit) in shared_commits_main.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see: main_commit3 (update), commit1 (initial)
-        assert_eq!(shared_commits_main.len(), 2);
-        assert_eq!(shared_commits_main[0].id, main_commit3);
-        assert_eq!(shared_commits_main[1].id, commit1);
-
-        // From feature branch
-        store.checkout("feature/new-keys").unwrap();
-        let shared_commits_feature = store.get_commits(b"shared_key").unwrap();
-        println!(
-            "Shared_key commits from feature branch: {}",
-            shared_commits_feature.len()
-        );
-        for (i, commit) in shared_commits_feature.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should see: branch_commit1 (update), commit1 (initial)
-        assert_eq!(shared_commits_feature.len(), 2);
-        assert_eq!(shared_commits_feature[0].id, branch_commit1);
-        assert_eq!(shared_commits_feature[1].id, commit1);
-
-        println!("\n=== Multi-branch commit tracking test completed successfully ===");
-    }
-
-    #[test]
-    fn test_get_commits_merge_scenarios() {
-        let temp_dir = TempDir::new().unwrap();
-        gix::init(temp_dir.path()).unwrap();
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // === Main branch setup ===
-        store
-            .insert(b"file1".to_vec(), b"main_content".to_vec())
-            .unwrap();
-        store
-            .insert(b"shared_file".to_vec(), b"original".to_vec())
-            .unwrap();
-        let main_commit1 = store.commit("Main: initial files").unwrap();
-
-        // === Feature branch development ===
-        store.create_branch("feature/enhancement").unwrap();
-        store.checkout("feature/enhancement").unwrap();
-
-        // Feature work
-        store
-            .insert(b"new_feature".to_vec(), b"feature_code".to_vec())
-            .unwrap();
-        store
-            .update(b"shared_file".to_vec(), b"feature_modified".to_vec())
-            .unwrap();
-        let feature_commit1 = store
-            .commit("Feature: add new feature and modify shared file")
-            .unwrap();
-
-        store
-            .update(b"new_feature".to_vec(), b"enhanced_feature_code".to_vec())
-            .unwrap();
-        let feature_commit2 = store.commit("Feature: enhance the new feature").unwrap();
-
-        // === Main branch continues ===
-        store.checkout("main").unwrap();
-
-        store
-            .update(b"file1".to_vec(), b"main_updated_content".to_vec())
-            .unwrap();
-        store
-            .insert(b"main_only".to_vec(), b"main_specific".to_vec())
-            .unwrap();
-        let main_commit2 = store
-            .commit("Main: update file1 and add main-specific file")
-            .unwrap();
-
-        // === Test commits before any merging ===
-        println!("\n=== Testing commits before merge ===");
-
-        // Test new_feature commits (should only exist on feature branch)
-        let feature_commits_from_main = store.get_commits(b"new_feature").unwrap();
-        assert_eq!(
-            feature_commits_from_main.len(),
-            0,
-            "new_feature should not exist on main branch"
-        );
-
-        store.checkout("feature/enhancement").unwrap();
-        let feature_commits_from_feature = store.get_commits(b"new_feature").unwrap();
-        assert_eq!(
-            feature_commits_from_feature.len(),
-            2,
-            "new_feature should have 2 commits on feature branch"
-        );
-        assert_eq!(feature_commits_from_feature[0].id, feature_commit2);
-        assert_eq!(feature_commits_from_feature[1].id, feature_commit1);
-
-        // Test shared_file evolution on different branches
-        let shared_commits_feature = store.get_commits(b"shared_file").unwrap();
-        assert_eq!(shared_commits_feature.len(), 2);
-        assert_eq!(shared_commits_feature[0].id, feature_commit1); // feature modification
-        assert_eq!(shared_commits_feature[1].id, main_commit1); // original
-
-        store.checkout("main").unwrap();
-        let shared_commits_main = store.get_commits(b"shared_file").unwrap();
-        assert_eq!(shared_commits_main.len(), 1);
-        assert_eq!(shared_commits_main[0].id, main_commit1); // only original on main
-
-        // === Test file1 commits (different evolution paths) ===
-        let file1_commits_main = store.get_commits(b"file1").unwrap();
-        assert_eq!(file1_commits_main.len(), 2);
-        assert_eq!(file1_commits_main[0].id, main_commit2); // main update
-        assert_eq!(file1_commits_main[1].id, main_commit1); // original
-
-        store.checkout("feature/enhancement").unwrap();
-        let file1_commits_feature = store.get_commits(b"file1").unwrap();
-        assert_eq!(file1_commits_feature.len(), 1);
-        assert_eq!(file1_commits_feature[0].id, main_commit1); // only original, no feature changes
-
-        println!("=== Merge scenario commit tracking test completed successfully ===");
-    }
-
-    #[test]
-    fn test_get_commits_key_lifecycle_patterns() {
-        let temp_dir = TempDir::new().unwrap();
-        gix::init(temp_dir.path()).unwrap();
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // === Pattern 1: Key created, modified multiple times, then deleted ===
-        store
-            .insert(b"lifecycle_key".to_vec(), b"v1".to_vec())
-            .unwrap();
-        let create_commit = store.commit("Create lifecycle_key").unwrap();
-
-        store
-            .update(b"lifecycle_key".to_vec(), b"v2".to_vec())
-            .unwrap();
-        let update1_commit = store.commit("Update lifecycle_key to v2").unwrap();
-
-        store
-            .update(b"lifecycle_key".to_vec(), b"v3".to_vec())
-            .unwrap();
-        let update2_commit = store.commit("Update lifecycle_key to v3").unwrap();
-
-        store
-            .update(b"lifecycle_key".to_vec(), b"v4_final".to_vec())
-            .unwrap();
-        let update3_commit = store.commit("Final update of lifecycle_key").unwrap();
-
-        store.delete(b"lifecycle_key").unwrap();
-        let delete_commit = store.commit("Delete lifecycle_key").unwrap();
-
-        // Test complete lifecycle
-        let lifecycle_commits = store.get_commits(b"lifecycle_key").unwrap();
-        println!("Lifecycle key commits: {}", lifecycle_commits.len());
-        for (i, commit) in lifecycle_commits.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        assert_eq!(lifecycle_commits.len(), 5);
-        assert_eq!(lifecycle_commits[0].id, delete_commit); // Most recent: deletion
-        assert_eq!(lifecycle_commits[1].id, update3_commit); // Final update
-        assert_eq!(lifecycle_commits[2].id, update2_commit); // v3 update
-        assert_eq!(lifecycle_commits[3].id, update1_commit); // v2 update
-        assert_eq!(lifecycle_commits[4].id, create_commit); // Original creation
-
-        // === Pattern 2: Key deleted and recreated ===
-        store
-            .insert(b"recreated_key".to_vec(), b"first_life".to_vec())
-            .unwrap();
-        let first_create = store.commit("First creation of recreated_key").unwrap();
-
-        store
-            .update(b"recreated_key".to_vec(), b"first_life_updated".to_vec())
-            .unwrap();
-        let first_update = store.commit("Update recreated_key in first life").unwrap();
-
-        store.delete(b"recreated_key").unwrap();
-        let first_delete = store.commit("Delete recreated_key").unwrap();
-
-        // Key is gone, let's add some other commits
-        store
-            .insert(b"other_key".to_vec(), b"other_value".to_vec())
-            .unwrap();
-        let _other_commit = store.commit("Add some other key").unwrap();
-
-        // Recreate the key
-        store
-            .insert(b"recreated_key".to_vec(), b"second_life".to_vec())
-            .unwrap();
-        let second_create = store
-            .commit("Recreate recreated_key with new value")
-            .unwrap();
-
-        store
-            .update(b"recreated_key".to_vec(), b"second_life_updated".to_vec())
-            .unwrap();
-        let second_update = store.commit("Update recreated_key in second life").unwrap();
-
-        // Test recreated key history
-        let recreated_commits = store.get_commits(b"recreated_key").unwrap();
-        println!("Recreated key commits: {}", recreated_commits.len());
-        for (i, commit) in recreated_commits.iter().enumerate() {
-            println!("  {}: {} - {}", i, commit.id, commit.message);
-        }
-
-        // Should track complete history including deletion and recreation
-        assert_eq!(recreated_commits.len(), 5);
-        assert_eq!(recreated_commits[0].id, second_update); // Latest update
-        assert_eq!(recreated_commits[1].id, second_create); // Recreation
-        assert_eq!(recreated_commits[2].id, first_delete); // Deletion
-        assert_eq!(recreated_commits[3].id, first_update); // Update in first life
-        assert_eq!(recreated_commits[4].id, first_create); // Original creation
-
-        // === Pattern 3: Key with no changes (single commit) ===
-        store
-            .insert(b"static_key".to_vec(), b"never_changes".to_vec())
-            .unwrap();
-        let static_commit = store.commit("Add static key that never changes").unwrap();
-
-        // Add other keys and commits
-        store
-            .insert(b"dynamic_key".to_vec(), b"changes_a_lot".to_vec())
-            .unwrap();
-        store.commit("Add dynamic key").unwrap();
-        store
-            .update(b"dynamic_key".to_vec(), b"changed_once".to_vec())
-            .unwrap();
-        store.commit("Update dynamic key").unwrap();
-        store
-            .update(b"dynamic_key".to_vec(), b"changed_again".to_vec())
-            .unwrap();
-        store.commit("Update dynamic key again").unwrap();
-
-        // Test static key (should only have one commit)
-        let static_commits = store.get_commits(b"static_key").unwrap();
-        println!("Static key commits: {}", static_commits.len());
-        assert_eq!(static_commits.len(), 1);
-        assert_eq!(static_commits[0].id, static_commit);
-
-        println!("=== Key lifecycle patterns test completed successfully ===");
-    }
-
-    #[test]
-    fn test_get_commits_empty_and_edge_cases() {
-        let temp_dir = TempDir::new().unwrap();
-        gix::init(temp_dir.path()).unwrap();
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir_all(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // === Test 1: Non-existent key ===
-        let nonexistent_commits = store.get_commits(b"does_not_exist").unwrap();
-        assert_eq!(
-            nonexistent_commits.len(),
-            0,
-            "Non-existent key should have no commits"
-        );
-
-        // === Test 2: Empty repository (no commits yet) ===
-        // This test happens before we make any commits
-        store
-            .insert(b"test_key".to_vec(), b"test_value".to_vec())
-            .unwrap();
-        // Don't commit yet - test with staged changes
-        let no_commits_yet = store.get_commits(b"test_key").unwrap();
-        assert_eq!(
-            no_commits_yet.len(),
-            0,
-            "Staged but uncommitted changes should show no commits"
-        );
-
-        // === Test 3: Make first commit ===
-        let first_commit = store.commit("First commit ever").unwrap();
-        let after_first_commit = store.get_commits(b"test_key").unwrap();
-        assert_eq!(after_first_commit.len(), 1);
-        assert_eq!(after_first_commit[0].id, first_commit);
-
-        // === Test 4: Key with empty value ===
-        store.insert(b"empty_key".to_vec(), vec![]).unwrap();
-        let empty_value_commit = store.commit("Add key with empty value").unwrap();
-
-        let empty_key_commits = store.get_commits(b"empty_key").unwrap();
-        assert_eq!(empty_key_commits.len(), 1);
-        assert_eq!(empty_key_commits[0].id, empty_value_commit);
-
-        // === Test 5: Key updated to empty value ===
-        store
-            .insert(b"becomes_empty".to_vec(), b"has_content".to_vec())
-            .unwrap();
-        let content_commit = store.commit("Add key with content").unwrap();
-
-        store.update(b"becomes_empty".to_vec(), vec![]).unwrap();
-        let empty_update_commit = store.commit("Update key to empty value").unwrap();
-
-        let empty_update_commits = store.get_commits(b"becomes_empty").unwrap();
-        assert_eq!(empty_update_commits.len(), 2);
-        assert_eq!(empty_update_commits[0].id, empty_update_commit);
-        assert_eq!(empty_update_commits[1].id, content_commit);
-
-        // === Test 6: Binary key and value ===
-        let binary_key = vec![0x00, 0x01, 0x02, 0xFF, 0xFE];
-        let binary_value = vec![0xDE, 0xAD, 0xBE, 0xEF, 0x00, 0xFF];
-
-        store
-            .insert(binary_key.clone(), binary_value.clone())
-            .unwrap();
-        let binary_commit = store.commit("Add binary key-value pair").unwrap();
-
-        let binary_commits = store.get_commits(&binary_key).unwrap();
-        assert_eq!(binary_commits.len(), 1);
-        assert_eq!(binary_commits[0].id, binary_commit);
-
-        // === Test 7: Very long key name ===
-        let long_key = b"very_long_key_name_".repeat(50); // 1000 characters
-        store
-            .insert(long_key.clone(), b"short_value".to_vec())
-            .unwrap();
-        let long_key_commit = store.commit("Add very long key name").unwrap();
-
-        let long_key_commits = store.get_commits(&long_key).unwrap();
-        assert_eq!(long_key_commits.len(), 1);
-        assert_eq!(long_key_commits[0].id, long_key_commit);
-
-        println!("=== Edge cases test completed successfully ===");
-    }
-
-    #[test]
-    fn test_thread_safe_basic_operations() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize a git repository
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to initialize git repository");
-
-        // Create a subdirectory for the dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_dir).unwrap();
-
-        let store = ThreadSafeGitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Test basic operations
-        store.insert(b"key1".to_vec(), b"value1".to_vec()).unwrap();
-        assert_eq!(store.get(b"key1"), Some(b"value1".to_vec()));
-
-        // Commit changes
-        store.commit("Initial commit").unwrap();
-
-        // Update key
-        store.update(b"key1".to_vec(), b"value2".to_vec()).unwrap();
-        assert_eq!(store.get(b"key1"), Some(b"value2".to_vec()));
-    }
-
-    #[test]
-    fn test_thread_safe_concurrent_access() {
-        use std::sync::Arc;
-        use std::thread;
-
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize a git repository
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to initialize git repository");
-
-        // Create a subdirectory for the dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_dir).unwrap();
-
-        let store = Arc::new(ThreadSafeGitVersionedKvStore::<32>::init(&dataset_dir).unwrap());
-
-        // Test concurrent reads and writes
-        let handles: Vec<_> = (0..5)
-            .map(|i| {
-                let store_clone = Arc::clone(&store);
-                thread::spawn(move || {
-                    let key = format!("key{}", i).into_bytes();
-                    let value = format!("value{}", i).into_bytes();
-                    store_clone.insert(key.clone(), value.clone()).unwrap();
-                    assert_eq!(store_clone.get(&key), Some(value));
-                })
-            })
-            .collect();
-
-        // Wait for all threads to complete
-        for handle in handles {
-            handle.join().unwrap();
-        }
-
-        // Verify all keys were inserted
-        store.commit("Concurrent insertions").unwrap();
-        let keys = store.list_keys().unwrap();
-        assert_eq!(keys.len(), 5);
-    }
-
-    #[test]
-    fn test_versioned_kv_store_merge() {
-        use crate::diff::TakeSourceResolver;
-        use tempfile::TempDir;
-
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize a git repository
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to initialize git repository");
-
-        // Configure git user for commits
-        std::process::Command::new("git")
-            .args(["config", "user.name", "Test User"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to configure git user name");
-
-        std::process::Command::new("git")
-            .args(["config", "user.email", "test@example.com"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to configure git user email");
-
-        // Create a subdirectory for the dataset
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create initial commit on main branch
-        store
-            .insert(b"shared".to_vec(), b"original".to_vec())
-            .unwrap();
-        store
-            .insert(b"main_only".to_vec(), b"main_value".to_vec())
-            .unwrap();
-        store.commit("Initial commit on main").unwrap();
-
-        // Create feature branch
-        store.create_branch("feature").unwrap();
-
-        // Make changes on feature branch
-        store
-            .insert(b"shared".to_vec(), b"feature_value".to_vec())
-            .unwrap();
-        store
-            .insert(b"feature_only".to_vec(), b"feature_value".to_vec())
-            .unwrap();
-        store.commit("Feature branch changes").unwrap();
-
-        // Switch back to main
-        store.checkout("main").unwrap();
-
-        // Make different changes on main
-        store
-            .insert(b"shared".to_vec(), b"main_modified".to_vec())
-            .unwrap();
-        store
-            .insert(b"main_new".to_vec(), b"main_new_value".to_vec())
-            .unwrap();
-        store.commit("Main branch changes").unwrap();
-
-        // Test merge with conflict resolver (take source)
-        let resolver = TakeSourceResolver;
-        let merge_result = store.merge("feature", &resolver);
-
-        assert!(
-            merge_result.is_ok(),
-            "Merge should succeed with conflict resolver"
-        );
-        let _merge_commit_id = merge_result.unwrap();
-
-        // Test the merge results
-        assert_eq!(
-            store.get(b"shared"),
-            Some(b"feature_value".to_vec()),
-            "shared value should be from feature branch"
-        );
-        assert_eq!(
-            store.get(b"main_only"),
-            Some(b"main_value".to_vec()),
-            "main_only should remain"
-        );
-        assert_eq!(
-            store.get(b"feature_only"),
-            Some(b"feature_value".to_vec()),
-            "feature_only should be added"
-        );
-        assert_eq!(
-            store.get(b"main_new"),
-            Some(b"main_new_value".to_vec()),
-            "main_new should remain"
-        );
-
-        // Verify we're still on main branch
-        assert_eq!(store.current_branch(), "main");
-    }
-
-    #[test]
-    fn test_versioned_kv_store_merge_ignore_conflicts() {
-        use tempfile::TempDir;
-
-        let temp_dir = TempDir::new().unwrap();
-
-        // Initialize git repository with proper config
-        std::process::Command::new("git")
-            .args(["init"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to initialize git repository");
-
-        std::process::Command::new("git")
-            .args(["config", "user.name", "Test User"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to configure git user name");
-
-        std::process::Command::new("git")
-            .args(["config", "user.email", "test@example.com"])
-            .current_dir(&temp_dir)
-            .output()
-            .expect("Failed to configure git user email");
-
-        let dataset_dir = temp_dir.path().join("dataset");
-        std::fs::create_dir(&dataset_dir).unwrap();
-
-        let mut store = GitVersionedKvStore::<32>::init(&dataset_dir).unwrap();
-
-        // Create base state
-        store
-            .insert(b"key1".to_vec(), b"base_value".to_vec())
-            .unwrap();
-        store.commit("Base commit").unwrap();
-
-        // Create feature branch with changes
-        store.create_branch("feature").unwrap();
-        store
-            .insert(b"key1".to_vec(), b"feature_value".to_vec())
-            .unwrap();
-        store
-            .insert(b"key2".to_vec(), b"feature_only".to_vec())
-            .unwrap();
-        store.commit("Feature changes").unwrap();
-
-        // Switch to main and make conflicting changes
-        store.checkout("main").unwrap();
-        store
-            .insert(b"key1".to_vec(), b"main_value".to_vec())
-            .unwrap();
-        store.commit("Main changes").unwrap();
-
-        // Test merge ignore conflicts (should keep main value for conflicts)
-        let merge_result = store.merge_ignore_conflicts("feature");
-
-        assert!(
-            merge_result.is_ok(),
-            "Merge ignore conflicts should succeed"
-        );
-
-        // Should keep main value for conflicting key, but add non-conflicting key
-        assert_eq!(store.get(b"key1"), Some(b"main_value".to_vec())); // Keep main value
-        assert_eq!(store.get(b"key2"), Some(b"feature_only".to_vec())); // Add from feature
-    }
-}
